### 2023-03-07 #5232 @editorialbot
[PRE REVIEW]: Rosalution: Supporting data accessibility, integration, curation, interoperability, and reuse for precision animal modeling
**Submitting author:** @serioushorncat (<a href="http://orcid.org/0000-0002-4631-9135">Angelina Uno-Antonison</a>)
**Repository:** https://github.com/uab-cgds-worthey/rosalution
**Branch with paper.md** (empty if default branch):
**Version:** 0.6.0
