### 2023-03-08 888baf0
updating to get the prop not the value, and lint fixs
