### 2023-03-08 183c824
Updated the script to remove the hard coded mongo docker container and added host parameters for mongofiles
