### 2023-03-08 1eddb36
Account view credentials (#46)
* Update user base model to have the client id and client secret

* Update AccountView.vue

* working on making clickable

* making progress on the client ID and client Secret

* updating how the secret is shown

* linting fixes

* refactored to have secretValue be a computed value

* working on unit tests

* working on the unit test.

* Delete UserInfoView.vue

* making progress on the unit tests

* working and passing unit tests

* Update frontend/src/views/AccountView.vue

Co-authored-by: Angelina Uno-Antonison <ange.unoantonison@gmail.com>
Signed-off-by: Alex
[…]
