### 2023-03-08 eff24f1
added the database variable to replace the hard-coded rosalution_db in the commands
