### 2023-03-09 30e6d03
More remove analysis script (#54)
* Updated the script to remove the hard coded mongo docker container and added host parameters for mongofiles

* added the database variable to replace the hard-coded rosalution_db in the commands

* Removed the echo for pedigree_id
