### 2023-03-09 bae2d63
1st draft for scripts to update the annotation configuration and create a backup.
