### 2023-03-10 38aff36
Script will create a backup of the annotation_config collection and replace the existing collection with the one provided to the script.
