### 2023-03-10 1107793
Cleaned up the script with improvements to the documentation and conslidated variables
