### 2023-03-10 24690b3
Resolved static analysis error SC2317 in regards to the combination of exit and the usage function exiting prematurely
