### 2023-03-10 7998a67
Updated missing fix for the name at the top of the script in the docs and added readme usage
