### 2023-03-13 a8cd43b
Apply suggestions from code review
Co-authored-by: James Scherer <jscherer@uab.edu>
Signed-off-by: Angelina Uno-Antonison <angelina.antonison@gmail.com>
