### 2023-03-13 aab04eb
Resolved the issue that was discoered regarding to always the script accidently always expecting to use docker; fixed backup default path information to be what it ought to be with the new backup directory
