+++ etc/database/backup-database.sh
@@ -4,8 +4,8 @@
 usage() {
   echo "usage: $0 <docker-container> <output-path (optional)>"
   echo " "
-  echo "Dumps a Rosalution MongoDB database archive named 'annotation-config-<current-date-seconds>.archive'"
-  echo "to the designated output path. By default, the output path is '/home/\$USER/backups/rosalution-db-backup/'"
+  echo "Dumps a Rosalution MongoDB database archive named 'rosalution-db-<current-date-seconds>.archive'"
+  echo "to the designated output path. By default, the output path is '/home/centos/backups/rosalution-db-backup/'"
   echo "Include a second paramater to overwrite the <output-path>.  If the <output-path> does not exist"
   echo "the backup operation will be canceled."
 }
@@ -20,7 +20,7 @@ then
 fi
 
 DOCKER_CONTAINER=$1
-TARGET_PATH=/home/$USER/backups/rosalution-db-backup/
+TARGET_PATH=/home/centos/backups/rosalution-db-backup/
 
 if [[ $# -eq 2 ]]
 then
