### 2023-03-13 796b93d
Annotation configuration update scripts (#55)
* 1st draft for scripts to update the annotation configuration and create a backup.
* Script will create a backup of the annotation_config collection and replace the existing collection with the one provided to the script.
* Cleaned up the script with improvements to the documentation and conslidated variables
* Resolved static analysis error SC2317 in regards to the combination of exit and the usage function exiting prematurely
* Updated missing fix for the name at the top of the script in the docs and added readme usage
* Apply suggestions from c
[…]
