### 2023-03-14 0e00142
added a component test for the new component CredentialsBox
