### 2023-03-14 1d41d6b
refactored unit test for the new component
