### 2023-03-15 c1f00bc
refactored to not require reloading the page on secret generation or update
