### 2023-03-15 4e3ff19
reorganizing code to make it easier to read.
