### 2023-03-15 1dbc1db
Create new UserInfoBox component
