### 2023-03-16 321e0ca
update the unit test to use Chai BDD syntax
