### 2023-03-22 efce8c3
Updated the pre-requesite documentation to include additional setup information that was discovered after asssiting some future contributors in setting up their development environment for rosalution
