### 2023-03-22 5825a64
updated to be more clear what are the dependencies with each of the tools needed for teh application.
