### 2023-03-22 e365ff7
Revise Rosalution prerequisite documentation to indicate which parts of Node.JS and Python  are needed (#60)
* Updated the pre-requisite documentation to include additional setup information that was discovered after assisting some future contributors in setting up their development environment for rosalution
