### 2023-03-23 7251417
Premilinary work on updating the annotation sections to handle multiple images. The images are now being handled by a new ImageDataset component and each section has a rendering configuration defined to handle images
