+++ frontend/src/models/analyses.js
@@ -25,6 +25,18 @@ export default {
     return await Requests.put(url, updatedSections);
   },
 
+  async markAnalysisReady(analysisName) {
+    const url = `/rosalution/api/analysis/${analysisName}/mark_ready`;
+    return await Requests.put(url);
+  },
+
+  async markAnalysisActive(analysisName) {
+    // const url = `/rosalution/api/analysis/${analysisName}/mark_active`;
+    // return await Requests.put(url);
+    console.warn('Mark analysis active does not exist yet');
+    //  ** Not implemented yet **
+  },
+
   async getAnnotationConfiguration(analysisName) {
     // const baseUrl = '/rosalution/api/';
     // const urlQuery = `analysis/{analysisName}`;
