### 2023-03-27 84cc53a
Added Mark Ready and Mark Active conditions to analysis actions, created unit test for Mark Active and Mark Ready, but of note Mark Ready endpoint does not currently exist
