### 2023-03-28 a830914
The multi-image attachment experience is working as expected
