### 2023-03-28 ecbf51c
UserInfoBox component (#59)
* fixed a typo

* changed to credentialsBox

* making changes to start to use credentialsBox component

* added the button and hard coded client ID & Secret

* fixed the button to display properly

* added a component test for the new component CredentialsBox

* refactored the on toggle and fixed some variables to have a standard naming convention

* refactored unit test for the new component

* fixed a typo

* refactored to not require reloading the page on secret generation or update

* lint fix

* reorganizing code to make it easier to read.

* removed console lo
[…]
