### 2023-03-29 56f66f7
Need to include the pull_request event trigger, so that outside collaborators can have GitHub Actions get approval to run on their pull request additions into Rosalution executed.
