+++ .github/workflows/python.yml
@@ -1,6 +1,6 @@
 name: Python CI
 
-on: [push, workflow_dispatch]
+on: [push, pull_request, workflow_dispatch]
 
 env:
   PYLINT_CONFIG: .pylintrc
