### 2023-03-29 fb36065
Need to include the pull_request event trigger, so that outside collaborators can have GitHub Actions get approval to run on their pull request additions into Rosalution executed. (#62)
