+++ .github/workflows/system-tests.yml
@@ -1,6 +1,6 @@
 name: System Tests
 
-on: [push, workflow_dispatch]
+on: [push, pull_request, workflow_dispatch]
 
 jobs:
   system-tests:
