+++ .github/workflows/nodejs.yml
@@ -1,6 +1,6 @@
 name: Node.js CI
 
-on: [push, workflow_dispatch]
+on: [push, pull_request, workflow_dispatch]
 
 defaults:
   run:
