### 2023-03-29 fd8e3c8
Apply suggestions from code review
Co-authored-by: James Scherer <james.m.scherer@gmail.com>
Signed-off-by: jbarkley256 <jbarkley256@gmail.com>
