### 2023-03-30 e68c130
Added a backend endpoint for removing annotation images, added a genomic_unit collection method to remove the annotation file, updated the model on the frontend to interact with it. Everything inbetween to make it work
