### 2023-03-30 ba31d49
update backend to have hardcoded third party links
these will change to be data driven instead of hardcoded in the future
