### 2023-03-30 a247c0b
update backend to have hardcoded third party links
these will change to be data driven instead of hardcoded in the future
