### 2023-03-31 82e1e6d
added monday & phenotips link attachment to the menu actions and added unit tests
