### 2023-04-03 a93c677
Working annotation image update
