### 2023-04-04 de4e148
Update image is working, cleaned up code and removed logs
