+++ frontend/src/fileRequests.js
@@ -1,8 +1,8 @@
 import Requests from '@/requests.js';
 
 export default {
-    async getImage(fileId) {
-        const url = `/rosalution/api/analysis/download/${fileId}`;
-        return await Requests.getImage(url);
-    }
-}
+  async getImage(fileId) {
+    const url = `/rosalution/api/analysis/download/${fileId}`;
+    return await Requests.getImage(url);
+  },
+};
