+++ backend/src/repository/genomic_unit_collection.py
@@ -85,9 +85,7 @@ class GenomicUnitCollection:
 
     def update_genomic_unit_with_mongo_id(self, genomic_unit_document):
         """ Takes a genomic unit and overwrites the existing object based on the object's id """
-        print("am I getting here?")
         genomic_unit_id = genomic_unit_document['_id']
-        print(genomic_unit_id)
         self.collection.update_one(
             {'_id': ObjectId(str(genomic_unit_id))},
             {'$set': genomic_unit_document},
