+++ frontend/src/models/annotations.js
@@ -33,7 +33,12 @@ export default {
       'section_name': annotation.section,
     };
 
-    return await Requests.postForm(`${baseUrl}/${annotation.genomic_unit}/attach/image`, attachmentForm);
+    const result = await Requests.postForm(`${baseUrl}/${annotation.genomic_unit}/attach/image`, attachmentForm);
+
+    console.log("Is this happening?")
+    console.log(result)
+
+    return result
   },
   async updateAnnotationImage(fileId, annotation, image) {
     const baseUrl = '/rosalution/api/annotate';
