+++ frontend/src/models/analyses.js
@@ -334,18 +334,15 @@ const annotationRenderingTemporary = [
     'header': 'Protein Expression',
     'anchor': 'Protein_Expression',
     'header_datasets': [],
-    'allowHeaderAttachImageDataset': true,
+    'allowHeaderAttachGenomicUnit': 'hgvs_variant',
     'rows': [{
       'class': '',
       'datasets': [
         {
           'dataset': 'Protein Expression',
           'type': 'image-dataset',
-          'props': {
-            'label': 'hgvs_variant'
-          }
-        }
-      ]
+        },
+      ],
     }],
   }, 
   {
@@ -354,7 +351,7 @@ const annotationRenderingTemporary = [
     'header': 'Modelability',
     'anchor': 'Modelability',
     'header_datasets': [],
-    'allowHeaderAttachImageDataset': true,
+    'allowHeaderAttachGenomicUnit': 'gene',
     'rows': [{
       'class': '',
       'datasets': [
