### 2023-04-06 65bb130
added a get summary by name endpoint & added tests
