+++ backend/tests/integration/test_analysis_routers.py
@@ -41,6 +41,16 @@ def test_get_analysis_summary(client, mock_access_token, mock_repositories):
     response = client.get("/analysis/summary", headers={"Authorization": "Bearer " + mock_access_token})
     assert len(response.json()) == 5
 
+def test_get_summary_by_name(client, mock_access_token, mock_repositories):
+    """Tests the summary_by_name endpoint"""
+    mock_repositories['analysis'].collection.find_one.return_value = read_test_fixture(
+        "analysis-CPAM0002.json")
+
+    response = client.get("/analysis/summary/CPAM0002", headers={"Authorization": "Bearer " + mock_access_token})
+
+    assert response.status_code == 200
+    assert response.json()["name"] == "CPAM0002"
+
 
 def test_import_analysis_from_phenotips_json(
     client,
