### 2023-04-07 dda7444
GitHub Action to deploy a successful compile & publish of production rosalution docker images to the devops cluster
