+++ .github/workflows/development-main-deployment.yml
@@ -1,0 +1,17 @@
+name: Production Build Deployment to DevOps Server
+
+on:
+  workflow_run:
+    workflows: [Compile & Publish]
+    types:
+      - completed
+  workflow_dispatch:
+
+env:
+  ROSALUTION_VERSION: ${{ vars.VERSION_BUILD_TAG }}
+
+jobs:
+  docker-shell-markdown-static-analysis:
+    runs-on: self-hosted
+      - name: Deploy Production Rosalution Build to Developer DevOsp Server
+        run: docker stack deploy --prune --with-registry-auth --compose-file docker-compose.production.yml rosalution-prod
