+++ .github/workflows/development-main-deployment.yml
@@ -13,5 +13,6 @@ env:
 jobs:
   docker-shell-markdown-static-analysis:
     runs-on: self-hosted
+    steps:
       - name: Deploy Production Rosalution Build to Developer DevOsp Server
         run: docker stack deploy --prune --with-registry-auth --compose-file docker-compose.production.yml rosalution-prod
