+++ .github/workflows/development-main-deployment.yml
@@ -6,6 +6,7 @@ on:
     types:
       - completed
   workflow_dispatch:
+  push:
 
 env:
   ROSALUTION_VERSION: ${{ vars.VERSION_BUILD_TAG }}
