+++ frontend/test/components/AnalysisView/AnalysisViewHeader.spec.js
@@ -14,6 +14,8 @@ function getMountedComponent(props) {
     titleText: 'CPAM0002',
     sectionAnchors: ['Brief', 'Summary', 'Medical'],
     actions: ['Action1', 'Action2', 'Action3'],
+    mondayLink: 'https://monday.com',
+    phenotipsLink: 'https://phenotips.org',
   };
 
   return shallowMount(AnalysisViewHeader, {
@@ -54,5 +56,32 @@ describe('AnalysisViewHeader.vue', () => {
     for (const expectedAnchorHref of expected) {
       expect(actualHrefs).to.include(expectedAnchorHref);
     }
+  });
+
+  it('should render third party links', () => {
+    const wrapper = getMountedComponent();
+
+    const mondayLink = wrapper.find('[data-test="monday-link"]');
+    expect(mondayLink.exists()).toBe(true);
+    expect(mondayLink.attributes('href')).to.equal('https://monday.com');
+    expect(mondayLink.attributes('target')).to.equal('_blank');
+
+    const phenotipsLink = wrapper.find('[data-test="phenotips-link"]');
+    expect(phenotipsLink.exists()).toBe(true);
+    expect(phenotipsLink.attributes('href')).to.equal('https://phenotips.org');
+    expect(phenotipsLink.attributes('target')).to.equal('_blank');
+  });
+
+  it('should not render third party links if null or empty string', () => {
+    const wrapper = getMountedComponent({
+      mondayLink: null,
+      phenotipsLink: '',
+    });
+
+    const mondayLink = wrapper.find('[data-test="monday-link"]');
+    expect(mondayLink.exists()).toBe(false);
+
+    const phenotipsLink = wrapper.find('[data-test="phenotips-link"]');
+    expect(phenotipsLink.exists()).toBe(false);
   });
 });
