+++ .github/workflows/development-main-deployment.yml
@@ -15,5 +15,10 @@ jobs:
   docker-shell-markdown-static-analysis:
     runs-on: self-hosted
     steps:
+      - name: Checkout repository
+        uses: actions/checkout@v3
+        with:
+          ref: HEAD^
+
       - name: Deploy Production Rosalution Build to Developer DevOsp Server
         run: docker stack deploy --prune --with-registry-auth --compose-file docker-compose.production.yml rosalution-prod
