### 2023-04-07 b86c138
Fixed the job name for production build deployment and removed the extra push command that was being used to test the workflow and github selfhosted runner
