+++ .github/workflows/build-and-publish-docker-images.yml
@@ -5,7 +5,6 @@ on:
     branches:
     - main
 
-
 env:
   REGISTRY: ghcr.io
   IMAGE_NAME: ${{ github.repository }}
