### 2023-04-07 b8c409a
Apply suggestions from code review
Co-authored-by: James Scherer <jscherer@uab.edu>
Signed-off-by: Angelina Uno-Antonison <angelina.antonison@gmail.com>
