### 2023-04-07 dfc82fe
adding logic and styling for adding 3rd party link logos to the cards
