### 2023-04-11 ee22072
updating unit tests to follow the ChaiJS BDD standard
