+++ frontend/test/views/AnnotationView.spec.js
@@ -80,8 +80,6 @@ describe('AnnotationView', () => {
     mockSummaryByName.returns(mockSummaryByNameForCPAM0002);
     
     annotationAttachMock = sandbox.stub(Annotations, 'attachAnnotationImage');
-
-    annotationAttachMock = sandbox.stub(Annotations, 'attachAnnotationImage');
     annotationUpdateMock = sandbox.stub(Annotations, 'updateAnnotationImage');
     annotationRemoveMock = sandbox.stub(Annotations, 'removeAnnotationImage');
 
@@ -162,22 +160,20 @@ describe('AnnotationView', () => {
           'https://www.ncbi.nlm.nih.gov/gene?Db=gene&Cmd=DetailsSearch&Term=ENSG00000160131',
         ]).to.include(linkDomElement.attributes('href'));
       });
-    });    
+    });
   });
 
   describe('Annotation image attachments', () => {
     describe('when an image section does not have an image', () => {
-      it.only('accepts an image to be added as content', async () => {
+      it('accepts an image to be added as content', async () => {
         const newImageResult = {
           image_id: 'fake-image-id-1',
           section: 'Gene Homology/Multi-Sequence Alignment',
-        }
+        };
 
         annotationAttachMock.returns(newImageResult);
 
-        const annotationSection = wrapper.findComponent('[id=Gene_Homology]')
-
-        // console.log(annotationSection)
+        const annotationSection = wrapper.findComponent('[id=Gene_Homology]');
 
         annotationSection.vm.$emit('attach-image', 'Gene Homology/Multi-Sequence Alignment', 'hgvs_variant');
 
@@ -187,24 +183,190 @@ describe('AnnotationView', () => {
 
         inputDialog.confirmation(fakeImage);
 
-        await wrapper.vm.$nextTick();
-        await wrapper.vm.$nextTick();
+        // Needs to cycle through updating the props in the view and then additional
+        // ticks for vuejs to reactively update the supplemental component
         await wrapper.vm.$nextTick();
         await wrapper.vm.$nextTick();
 
-        const reRenderedAnnotationSection = wrapper.findComponent('[id=Gene_Homology]');
+        const tinyImageDatasetComponents = wrapper.findAllComponents(TinyImageDataset);
 
-        // console.log(wrapper.vm.annotations);
+        expect(tinyImageDatasetComponents.length).to.equal(1);
+      });
 
-        const imageDatasetComponents = reRenderedAnnotationSection.findComponent(ImageDataset)
+      it('should display a notification if the image fails to upload', async () => {
+        annotationAttachMock.throws('Something went wrong when attaching an image. Please seek help.');
 
-        // console.log(imageDatasetComponents.vm.value)
+        const annotationSection = wrapper.findComponent('[id=Gene_Homology]');
+        annotationSection.vm.$emit('attach-image', 'Gene Homology/Multi-Sequence Alignment', 'hgvs_variant');
+
+        await wrapper.vm.$nextTick();
+
+        const fakeImage = {data: 'path/to/fake/fakeImage.png'};
+        inputDialog.confirmation(fakeImage);
+
+        // Needs to cycle through updating the props in the view and then additional
+        // ticks for vuejs to reactively update the supplemental component
+        await wrapper.vm.$nextTick();
+        await wrapper.vm.$nextTick();
+
+        notificationDialog.confirmation();
+
+        await wrapper.vm.$nextTick();
+
+        const failureNotificationDialog = wrapper.findComponent(NotificationDialog);
+        expect(failureNotificationDialog.exists()).to.be.true;
       });
     });
 
     describe('when an image section has an image', () => {
[…]
