+++ frontend/test/models/annotations.spec.js
@@ -4,14 +4,16 @@ import Annotations from '@/models/annotations.js';
 import Requests from '@/requests.js';
 import sinon from 'sinon';
 
-describe('analyses.js', () => {
+describe('annotations.js', () => {
   const sandbox = sinon.createSandbox();
   let mockGetRequest;
   let mockPostFormRequest;
+  let mockDeleteFormRequest;
 
   beforeEach(() => {
     mockGetRequest = sandbox.stub(Requests, 'get');
     mockPostFormRequest = sandbox.stub(Requests, 'postForm');
+    mockDeleteFormRequest = sandbox.stub(Requests, 'deleteForm');
   });
 
   afterEach(() => {
@@ -59,5 +61,56 @@ describe('analyses.js', () => {
 
     expect(actualReturned).to.equal(expectedReturn);
     expect(mockPostFormRequest.calledWith(expectedUrl, expectedFormData)).to.be.true;
+  });
+
+  it('saves a new image over an existing image with its corresponding section name', async () => {
+    const expectedUrl = '/rosalution/api/annotate/SBFP1/update/image/fake-image-id-1';
+    const expectedFormData = {
+      'upload_file': 'file here',
+      'section_name': 'Gene Homology',
+      'genomic_unit_type': 'gene',
+    };
+
+    const expectedReturn = 'it worked';
+
+    mockPostFormRequest.returns(expectedReturn);
+
+    const actualReturned = await Annotations.updateAnnotationImage(
+        'fake-image-id-1',
+        {
+          genomic_unit: 'SBFP1',
+          section: 'Gene Homology',
+          genomic_unit_type: 'gene',
+        },
+        'file here',
+    );
+
+    expect(actualReturned).to.equal(expectedReturn);
+    expect(mockPostFormRequest.calledWith(expectedUrl, expectedFormData)).to.be.true;
+  });
+
+  it('removes an image annotation from its corresponding section name', async () => {
+    const expectedUrl = '/rosalution/api/annotate/SBFP1/remove/image/fake-image-id-1';
+    const expectedFormData = {
+      'genomic_unit_type': 'gene',
+      'section_name': 'Gene Homology',
+
+    };
+
+    const expectedReturn = 'it worked';
+
+    mockDeleteFormRequest.returns(expectedReturn);
+
+    const actualReturned = await Annotations.removeAnnotationImage(
+        'fake-image-id-1',
+        {
+          genomic_unit: 'SBFP1',
+          section: 'Gene Homology',
+          genomic_unit_type: 'gene',
+        },
+    );
+
+    expect(actualReturned).to.equal(expectedReturn);
+    expect(mockDeleteFormRequest.calledWith(expectedUrl, expectedFormData)).to.be.true;
   });
 });
