+++ frontend/test/views/AnnotationView.spec.js
@@ -4,16 +4,52 @@ import sinon from 'sinon';
 
 import Analyses from '@/models/analyses.js';
 import Annotations from '@/models/annotations.js';
+import FileRequests from '@/fileRequests.js';
 
 import AnnotationView from '@/views/AnnotationView.vue';
 import AnnotationSection from '@/components/AnnotationView/AnnotationSection.vue';
 import AnnotationViewHeader from '@/components/AnnotationView/AnnotationViewHeader.vue';
 import TextDataset from '@/components/AnnotationView/TextDataset.vue';
-import ImageDataset from '@/components/AnnotationView/ImageDataset.vue';
+import TinyImageDataset from '@/components/AnnotationView/TinyImageDataset.vue';
+
+import NotificationDialog from '@/components/Dialogs/NotificationDialog.vue';
 
 import inputDialog from '@/inputDialog.js';
+import notificationDialog from '@/notificationDialog.js';
 
 import {FontAwesomeIcon} from '@fortawesome/vue-fontawesome';
+
+/**
+ * Helper mounts and returns the rendered component
+ * @param {props} props props for testing to overwrite default props
+ * @return {VueWrapper} returns a shallow mounted using props
+ */
+function getMountedComponent(props) {
+  const defaultProps = {analysis_name: 'CPAM0046'};
+  const mockRoute = {
+    params: {
+      id: 1,
+    },
+  };
+  return mount(AnnotationView, {
+    props: {...defaultProps, ...props},
+    global: {
+      components: {
+        'font-awesome-icon': FontAwesomeIcon,
+      },
+      mocks: {
+        $route: mockRoute,
+        $router: {
+          push: sinon.spy(),
+        },
+      },
+      stubs: {
+        AnnotationViewHeader: true,
+        FontAwesomeIcon: true,
+      },
+    },
+  });
+}
 
 describe('AnnotationView', () => {
   let wrapper;
@@ -22,25 +58,16 @@ describe('AnnotationView', () => {
   // Future Mock Analysis Annotation Rendering Configuration
 
   let annotationAttachMock;
+  let annotationUpdateMock;
+  let annotationRemoveMock;
 
-  const mockRoute = {
-    params: {
-      id: 1,
-    },
-  };
+  let getImageMock;
 
-  const mockRouter = {
-    push: sinon.spy(),
-  };
   let sandbox;
 
   beforeAll(() => {
     config.global.renderStubDefaultSlot = true;
     sandbox = sinon.createSandbox();
-
-    const defaultProps = {
-      analysis_name: 'CPAM0046',
-    };
 
     mockAnnotations = sandbox.stub(Annotations, 'getAnnotations');
     mockAnnotations.returns(mockAnnotationsForCPAM0002);
@@ -49,23 +76,13 @@ describe('AnnotationView', () => {
     mockGenomicUnits.returns(mockGenomicUnitsForCPAM0002);
 
     annotationAttachMock = sandbox.stub(Annotations, 'attachAnnotationImage');
+    annotationUpdateMock = sandbox.stub(Annotations, 'updateAnnotationImage');
+    annotationRemoveMock = sandbox.stub(Annotations, 'removeAnnotationImage');
 
-    wrapper = mount(AnnotationView, {
-      props: {...defaultProps},
-      global: {
-        components: {
-          'font-awesome-icon': FontAwesomeIcon,
-        },
-        mocks: {
-          $route: mockRoute,
-          $router: mockRouter,
-        },
-        stubs: {
-          AnnotationViewHeader: true,
-          FontAwesomeIcon: true,
-        },
-      },
-    });
+    getImageMock = sandbox.stub(FileRequests, 'getImage');
[…]
