### 2023-04-12 e59a5d1
Updated ImageDataset to be ImagesDataset, Added try and catch statements for the annotation file routes
