+++ backend/src/repository/genomic_unit_collection.py
@@ -137,6 +137,7 @@ class GenomicUnitCollection:
 
         for annotation in genomic_unit_document['annotations']:
             if data_set in annotation:
+                print(annotation)
                 annotation[data_set][0]['value'].append(genomic_annotation['value'])
                 self.update_genomic_unit_with_mongo_id(genomic_unit_document)
                 return
@@ -159,8 +160,6 @@ class GenomicUnitCollection:
 
         genomic_unit_document = self.find_genomic_unit(genomic_unit)
         data_set = section_name
-
-        # print(genomic_annotation_value)
 
         for annotation in genomic_unit_document['annotations']:
             if data_set in annotation:
