### 2023-04-13 9317521
Created a migration script for the file annotations to ensure the images will work when this update goes in
