### 2023-04-14 2cce482
Updated the update-file-annotation.js to give help/usage details
