### 2023-04-14 1a4d79a
updated update-file-annotation.js -> migrate-individual-image-dataset-to-multi-image-dataset.js
