+++ etc/fixtures/migrate-individual-image-dataset-to-multi-image-dataset.js
@@ -1,5 +1,5 @@
 const usage = `
-mongosh /tmp/fixtures/add-user.js
+mongosh /tmp/fixtures/migrate-individual-image-dataset-to-multi-image-dataset.js
     Script Options:
         help: if true print this help message
     
@@ -7,9 +7,9 @@ mongosh /tmp/fixtures/add-user.js
     
     Example: 
     
-    mongosh --host localhost --port 27017 --eval "var help=true;" /tmp/fixtures/update-file-anntation.js
+    mongosh --host localhost --port 27017 --eval "var help=true;" /tmp/fixtures/migrate-individual-image-dataset-to-multi-image-dataset.js
     
-    docker exec -it <rosalution_db_container> mongosh /tmp/fixtures/update-file-anntation.js
+    docker exec -it <rosalution_db_container> mongosh /tmp/fixtures/migrate-individual-image-dataset-to-multi-image-dataset.js
 `
 
 if (help === true) {
