### 2023-04-18 eb3a007
Third party link logo in headers (#64)
* starting to be able to add Monday.com and PhenoTips to an analysis

* made this easier to read and fixed some typos

* lint fix

* lint fix python

* undid the changes to the analysis collection.

* add monday.com logo

* update backend to have hardcoded third party links

these will change to be data driven instead of hardcoded in the future

* started some work on the header portion.

will carry this over for the next PR after the next one for the menu

* changed to camel case

* contains a mix of work for the header and adding an third party link in 
[…]
