### 2023-04-19 0ea0cb2
making this much more data driven and robust
