### 2023-04-19 38b8a02
Update backend/tests/unit/repository/test_genomic_unit_collection.py
Co-authored-by: Angelina Uno-Antonison <ange.unoantonison@gmail.com>
Signed-off-by: James Scherer <james.m.scherer@gmail.com>
