### 2023-04-20 4534a80
Updated the routes to follow the format of /<genomic_unit>/<data_set/ to be more descriptive, this includes test changes and re-organizing how data is structured and passed
