+++ etc/fixtures/migrations/migrate-individual-image-dataset-to-multi-image-dataset.js
@@ -1,5 +1,5 @@
 const usage = `
-mongosh /tmp/fixtures/migrate-individual-image-dataset-to-multi-image-dataset.js
+mongosh /tmp/fixtures/migrations/migrate-individual-image-dataset-to-multi-image-dataset.js
     Script Options:
         help: if true print this help message
     
@@ -7,9 +7,9 @@ mongosh /tmp/fixtures/migrate-individual-image-dataset-to-multi-image-dataset.js
     
     Example: 
     
-    mongosh --host localhost --port 27017 --eval "var help=true;" /tmp/fixtures/migrate-individual-image-dataset-to-multi-image-dataset.js
+    mongosh --host localhost --port 27017 --eval "var help=true;" /tmp/fixtures/migrations/migrate-individual-image-dataset-to-multi-image-dataset.js
     
-    docker exec -it <rosalution_db_container> mongosh /tmp/fixtures/migrate-individual-image-dataset-to-multi-image-dataset.js
+    docker exec -it <rosalution_db_container> mongosh /tmp/fixtures/migrations/migrate-individual-image-dataset-to-multi-image-dataset.js
 `
 
 if (help === true) {
@@ -34,7 +34,7 @@ const annotationImageSections = [
 ]
 
 const oldGeneHomologyKey = 'Gene Homology/Multi-Sequence allignment';
-const newGeneHomologyKey = 'Gene Homology/Multi-Sequence Alignment';
+const newGeneHomologyKey = 'GeneHomology_Multi-SequenceAlignment';
 
 try {
     const genomicUnits = db.genomic_units.find();
