+++ backend/tests/integration/test_analysis_routers.py
@@ -400,25 +400,23 @@ def test_attach_third_party_link(client, mock_access_token, mock_repositories):
     """ Testing the attach third party link endpoint """
     mock_repositories["analysis"].collection.find_one.return_value = read_test_fixture("analysis-CPAM0002.json")
     expected = read_test_fixture("analysis-CPAM0002.json")
-    expected["third_party_links"] = [{"type": "MONDAY_COM", "link": "https://monday.com"}]
-    # print(expected)
+    expected["third_party_links"] = [{"type": "monday_com", "link": "https://monday.com"}]
     mock_repositories["analysis"].collection.find_one_and_update.return_value = expected
     response = client.put(
-        "/analysis/CPAM0002/attach/MONDAY_COM",
+        "/analysis/CPAM0002/attach/monday_com",
         headers={"Authorization": "Bearer " + mock_access_token},
         data={"link": "https://monday.com"}
     )
 
     assert response.status_code == 200
-    # print(response.json())
-    assert response.json()["third_party_links"] == [{"type": "MONDAY_COM", "link": "https://monday.com"}]
+    assert response.json()["third_party_links"] == [{"type": "monday_com", "link": "https://monday.com"}]
 
 
 def test_attach_third_party_link_analysis_does_not_exist(client, mock_access_token, mock_repositories):
     """ Testing the attach third party link endpoint """
     mock_repositories["analysis"].collection.find_one.return_value = None
     response = client.put(
-        "/analysis/CPAM0002/attach/MONDAY_COM",
+        "/analysis/CPAM0002/attach/monday_com",
         headers={"Authorization": "Bearer " + mock_access_token},
         data={"link": "monday.com"}
     )
@@ -435,7 +433,7 @@ def test_attach_third_party_link_invalid_enum(client, mock_access_token, mock_re
         data={"link": "monday.com"}
     )
 
-    assert response.status_code == 409
+    assert response.status_code == 422
 
 
 def test_mark_ready(client, mock_access_token, mock_repositories):
