+++ frontend/test/components/AnnotationView/AnnotationViewHeader.spec.js
@@ -24,10 +24,6 @@ function getMountedComponent(props) {
       'gene': 'PEX10',
       'variant': 'NM_153818.2:c.28dup(p.Glu10fs)',
     },
-    third_party_links: [
-      {'type': 'monday_com', 'link': 'https://monday.com'},
-      {'type': 'phenotips_com', 'link': 'https://phenotips.com'},
-    ],
   };
 
   return shallowMount(AnnotationViewHeader, {
@@ -114,28 +110,5 @@ describe('AnnotationViewHeader.vue', () => {
     geneSelectComponent.vm.$emit('update:selected', 'PEX10');
 
     expect(wrapper.emitted('changed'), 'the "changed" event happened').to.be.undefined;
-  });
-
-  it('should render third party links', () => {
-    const wrapper = getMountedComponent();
-
-    const thirdPartyLinks = wrapper.findAll('[data-test="third-party-link"]');
-
-    const mondayLink = thirdPartyLinks[0];
-    expect(mondayLink.attributes('href')).to.equal('https://monday.com');
-    expect(mondayLink.attributes('target')).to.equal('_blank');
-
-    const phenotipsLink = thirdPartyLinks[1];
-    expect(phenotipsLink.attributes('href')).to.equal('https://phenotips.com');
-    expect(phenotipsLink.attributes('target')).to.equal('_blank');
-  });
-
-  it('should not render third party links if empty', () => {
-    const wrapper = getMountedComponent({
-      third_party_links: [],
-    });
-
-    const thirdPartyLink = wrapper.find('[data-test="third-party-link"]');
-    expect(thirdPartyLink.exists()).to.be.false;
   });
 });
