+++ frontend/test/views/AnnotationView.spec.js
@@ -78,16 +78,13 @@ describe('AnnotationView', () => {
 
     mockSummaryByName = sandbox.stub(Analyses, 'getSummaryByName');
     mockSummaryByName.returns(mockSummaryByNameForCPAM0002);
-    
+
     annotationAttachMock = sandbox.stub(Annotations, 'attachAnnotationImage');
     annotationUpdateMock = sandbox.stub(Annotations, 'updateAnnotationImage');
     annotationRemoveMock = sandbox.stub(Annotations, 'removeAnnotationImage');
 
     getImageMock = sandbox.stub(FileRequests, 'getImage');
     getImageMock.returns('data:image/jpeg;base64,/fake-data');
-
-    mockSummaryByName = sandbox.stub(Analyses, 'getSummaryByName');
-    mockSummaryByName.returns(mockSummaryByNameForCPAM0002);
 
     wrapper = getMountedComponent();
   });
@@ -162,214 +159,6 @@ describe('AnnotationView', () => {
           'https://www.ncbi.nlm.nih.gov/gene?Db=gene&Cmd=DetailsSearch&Term=HGNC:22082',
           'https://www.ncbi.nlm.nih.gov/gene?Db=gene&Cmd=DetailsSearch&Term=ENSG00000160131',
         ]).to.include(linkDomElement.attributes('href'));
-      });
-    });
-  });
-
-  describe('Annotation image attachments', () => {
-    describe('when an image section does not have an image', () => {
-      it('accepts an image to be added as content', async () => {
-        const newImageResult = {
-          image_id: 'fake-image-id-1',
-          section: 'Gene Homology/Multi-Sequence Alignment',
-        };
-
-        annotationAttachMock.returns(newImageResult);
-
-        const annotationSection = wrapper.findComponent('[id=Gene_Homology]');
-
-        annotationSection.vm.$emit('attach-image', 'Gene Homology/Multi-Sequence Alignment', 'hgvs_variant');
-
-        await wrapper.vm.$nextTick();
-
-        const fakeImage = {data: 'path/to/fake/fakeImage.png'};
-
-        inputDialog.confirmation(fakeImage);
-
-        // Needs to cycle through updating the props in the view and then additional
-        // ticks for vuejs to reactively update the supplemental component
-        await wrapper.vm.$nextTick();
-        await wrapper.vm.$nextTick();
-
-        const tinyImageDatasetComponents = wrapper.findAllComponents(TinyImageDataset);
-
-        expect(tinyImageDatasetComponents.length).to.equal(1);
-      });
-
-      it('should display a notification if the image fails to upload', async () => {
-        annotationAttachMock.throws('Something went wrong when attaching an image. Please seek help.');
-
-        const annotationSection = wrapper.findComponent('[id=Gene_Homology]');
-        annotationSection.vm.$emit('attach-image', 'Gene Homology/Multi-Sequence Alignment', 'hgvs_variant');
-
-        await wrapper.vm.$nextTick();
-
-        const fakeImage = {data: 'path/to/fake/fakeImage.png'};
-        inputDialog.confirmation(fakeImage);
-
-        // Needs to cycle through updating the props in the view and then additional
-        // ticks for vuejs to reactively update the supplemental component
-        await wrapper.vm.$nextTick();
-        await wrapper.vm.$nextTick();
-
-        notificationDialog.confirmation();
-
-        await wrapper.vm.$nextTick();
-
-        const failureNotificationDialog = wrapper.findComponent(NotificationDialog);
-        expect(failureNotificationDialog.exists()).to.be.true;
-      });
-    });
-
-    describe('when an image section has an image', () => {
-      beforeEach(() => {
-        const imageAnnotation = {'file_id': 'fake-image-id-1', 'created_date': 'fake-date'};
-
[…]
