### 2023-04-21 9878dba
Updated Protein Expression to use gene consistently instead of hgvs_variant. This solves an issue with fetching image annotations
