+++ frontend/src/models/analyses.js
@@ -347,7 +347,7 @@ const annotationRenderingTemporary = [
   //       }
   //     ]
   //   }],
-  // }, 
+  // },
   // {
   //   'type': 'section',
   //   'class': '',
