### 2023-04-21 d905f00
Removed old rendering config
