### 2023-04-21 c124baa
Removed some console logs and general linting
