### 2023-04-21 abea2a5
As a computational biologist, I want to be able to attach multiple images to an annotation, so that I can more flexibly associate visual data (#67)
* Premilinary work on updating the annotation sections to handle multiple images. The images are now being handled by a new ImageDataset component and each section has a rendering configuration defined to handle images

* Work in progress. Multiple images work in the backend with uploading and fetching, but displaying them is the challenge

* The multi-image attachment experience is working as expected

* Looking into how to remove the multiple att
[…]
