+++ backend/tests/unit/repository/test_genomic_unit_collection.py
@@ -1,5 +1,6 @@
 """ Manages the genomic unit collection. Including reading, writing, fetching various genomic units. """
 from unittest.mock import Mock
+import copy
 import pytest
 
 from bson import ObjectId
@@ -225,6 +226,98 @@ def test_annotate_transcript_genomic_unit(genomic_unit_collection):
                 "annotations": {},
             }
         })
+
+
+def test_annotation_genomic_unit_with_file(genomic_unit_collection, hgvs_variant_genomic_unit_json):
+    """ Accepts a file and adds it as an annotation to the given genomic unit """
+    genomic_unit = {'unit': 'NM_001017980.3:c.164G>T', 'type': GenomicUnitType.HGVS_VARIANT}
+    annotation_unit = {
+        "data_set": 'GeneHomology_Multi-SequenceAlignment', "data_source": "rosalution_manual", "version": "1979-01-01",
+        "value": {"file_id": "fake-image-id-1", "created_date": "1979-01-01 00:00:00"}
+    }
+
+    expected_annotation_update = {
+        "GeneHomology_Multi-SequenceAlignment": [{
+            "data_source": "rosalution_maual", "version": "1979-01-01",
+            "value": [{"file_id": "fake-image-id-1", "created_date": "1979-01-01 00:00:00"}]
+        }]
+    }
+
+    expected_genomic_unit = hgvs_variant_genomic_unit_json
+    expected_genomic_unit['annotations'].append(expected_annotation_update)
+
+    genomic_unit_collection.collection.find_one.return_value = hgvs_variant_genomic_unit_json
+    genomic_unit_collection.update_genomic_unit_with_mongo_id = Mock()
+
+    genomic_unit_collection.annotate_genomic_unit_with_file(genomic_unit, annotation_unit)
+
+    genomic_unit_collection.update_genomic_unit_with_mongo_id.assert_called_once_with(expected_genomic_unit)
+
+
+def test_update_existing_genomic_unit_file_annotation(genomic_unit_collection, hgvs_variant_genomic_unit_json):
+    """ Updates an existing file annotation with a new file annotation in place"""
+    genomic_unit = {'unit': 'NM_001017980.3:c.164G>T', 'type': GenomicUnitType.HGVS_VARIANT}
+    annotation_unit_value = {"file_id": "fake-image-id-2", "created_date": "1979-01-01 00:00:00"}
+    data_set = 'GeneHomology_Multi-SequenceAlignment'
+    file_id_old = "fake-image-id-1"
+
+    expected_genomic_unit = copy.copy(hgvs_variant_genomic_unit_json)
+
+    hgvs_variant_genomic_unit_json['annotations'].append({
+        "GeneHomology_Multi-SequenceAlignment": [{
+            "data_source": "rosalution_maual", "version": "1979-01-01",
+            "value": [{"file_id": "fake-image-id-1", "created_date": "1979-01-01 00:00:00"}]
+        }]
+    })
+
+    genomic_unit_collection.collection.find_one.return_value = hgvs_variant_genomic_unit_json
+
+    expected_annotation_update = {
+        "GeneHomology_Multi-SequenceAlignment": [{
+            "data_source": "rosalution_maual", "version": "1979-01-01",
+            "value": [{"file_id": "fake-image-id-2", "created_date": "1979-01-01 00:00:00"}]
+        }]
+    }
+
+    expected_genomic_unit['annotations'].append(expected_annotation_update)
+
+    genomic_unit_collection.update_genomic_unit_with_mongo_id = Mock()
+
+    genomic_unit_collection.update_genomic_unit_file_annotation(
[…]

[…]
