+++ frontend/test/components/RosalutionHeader.spec.js
@@ -100,65 +100,65 @@ describe('RosalutionHeaderComponent.vue', () => {
   });
 
   it('should display the proper workflow icon based on the workflow_status prop', async () => {
-    const wrapper = getMountedComponent({ workflow_status: 'Annotation' });
-    let statusIcons = wrapper.findAllComponents({ name: 'FontAwesomeIcon' });
+    const wrapper = getMountedComponent({workflow_status: 'Annotation'});
+    let statusIcons = wrapper.findAllComponents({name: 'FontAwesomeIcon'});
     let statusIcon = statusIcons[0];
     expect(statusIcon.props('icon')).to.equal('asterisk');
 
-    await wrapper.setProps({ workflow_status: 'Ready' });
-    statusIcons = wrapper.findAllComponents({ name: 'FontAwesomeIcon' });
+    await wrapper.setProps({workflow_status: 'Ready'});
+    statusIcons = wrapper.findAllComponents({name: 'FontAwesomeIcon'});
     statusIcon = statusIcons[0];
     expect(statusIcon.props('icon')).to.equal('clipboard-check');
 
-    await wrapper.setProps({ workflow_status: 'Active' });
-    statusIcons = wrapper.findAllComponents({ name: 'FontAwesomeIcon' });
+    await wrapper.setProps({workflow_status: 'Active'});
+    statusIcons = wrapper.findAllComponents({name: 'FontAwesomeIcon'});
     statusIcon = statusIcons[0];
     expect(statusIcon.props('icon')).to.equal('book-open');
 
-    await wrapper.setProps({ workflow_status: 'Approved' });
-    statusIcons = wrapper.findAllComponents({ name: 'FontAwesomeIcon' });
+    await wrapper.setProps({workflow_status: 'Approved'});
+    statusIcons = wrapper.findAllComponents({name: 'FontAwesomeIcon'});
     statusIcon = statusIcons[0];
     expect(statusIcon.props('icon')).to.equal('check');
 
-    await wrapper.setProps({ workflow_status: 'On-Hold' });
-    statusIcons = wrapper.findAllComponents({ name: 'FontAwesomeIcon' });
+    await wrapper.setProps({workflow_status: 'On-Hold'});
+    statusIcons = wrapper.findAllComponents({name: 'FontAwesomeIcon'});
     statusIcon = statusIcons[0];
     expect(statusIcon.props('icon')).to.equal('pause');
 
-    await wrapper.setProps({ workflow_status: 'Declined' });
-    statusIcons = wrapper.findAllComponents({ name: 'FontAwesomeIcon' });
+    await wrapper.setProps({workflow_status: 'Declined'});
+    statusIcons = wrapper.findAllComponents({name: 'FontAwesomeIcon'});
     statusIcon = statusIcons[0];
     expect(statusIcon.props('icon')).to.equal('x');
   });
 
   it('should display the proper workflow color based on the workflow_status prop', async () => {
-    const wrapper = getMountedComponent({ workflow_status: 'Annotation' });
-    let statusIcons = wrapper.findAllComponents({ name: 'FontAwesomeIcon' });
+    const wrapper = getMountedComponent({workflow_status: 'Annotation'});
+    let statusIcons = wrapper.findAllComponents({name: 'FontAwesomeIcon'});
     let statusIcon = statusIcons[0];
     expect(statusIcon.attributes('style')).to.equal('color: var(--rosalution-status-annotation);');
 
-    await wrapper.setProps({ workflow_status: 'Ready' });
-    statusIcons = wrapper.findAllComponents({ name: 'FontAwesomeIcon' });
+    await wrapper.setProps({workflow_status: 'Ready'});
[…]
