### 2023-04-25 6f23f24
3r party array refactor (#70)
* backend refactor

* partial front end refactor

* making this much more data driven and robust

* front-end refactor

* lint fix

* backend refactor requested changes

* migrating where the logic for the third party link logos in the header

* passing front end unit tests

* python linting fix

* PR requested changes.

* removed the unnecessary computed value
