+++ backend/tests/unit/repository/test_analysis_collection.py
@@ -6,7 +6,6 @@ import pytest
 from pymongo import ReturnDocument
 from src.enums import EventType
 from src.models.event import Event
-from src.enums import ThirdPartyLinkType
 from ...test_utils import read_test_fixture
 
 
@@ -243,23 +242,21 @@ def test_remove_pedigree_file_id(analysis_collection):
 def test_attach_third_party_link_monday(analysis_collection):
     """Tests the attach_third_party_link function"""
     analysis_collection.collection.find_one.return_value = read_test_fixture("analysis-CPAM0002.json")
-    analysis_collection.attach_third_party_link("CPAM0002", "MONDAY_COM", "https://monday.com")
-    analysis_collection.collection.find_one_and_update.assert_called_with(
-        {'name': 'CPAM0002'},
-        {'$set': {ThirdPartyLinkType.MONDAY_COM: 'https://monday.com'}},
-        return_document=True,
-    )
+    analysis_collection.attach_third_party_link("CPAM0002", "monday_com", "https://monday.com")
+    analysis_collection.collection.find_one_and_update.assert_called_with({'name': 'CPAM0002'}, {
+        '$push': {'third_party_links': {'type': "monday_com", 'link': "https://monday.com"}}
+    },
+                                                                          return_document=True)
 
 
 def test_attach_third_party_link_phenotips(analysis_collection):
     """Tests the attach_third_party_link function"""
     analysis_collection.collection.find_one.return_value = read_test_fixture("analysis-CPAM0002.json")
-    analysis_collection.attach_third_party_link("CPAM0002", "PHENOTIPS_COM", "https://phenotips.com")
-    analysis_collection.collection.find_one_and_update.assert_called_with(
-        {'name': 'CPAM0002'},
-        {'$set': {ThirdPartyLinkType.PHENOTIPS_COM: 'https://phenotips.com'}},
-        return_document=True,
-    )
+    analysis_collection.attach_third_party_link("CPAM0002", "phenotips_com", "https://phenotips.com")
+    analysis_collection.collection.find_one_and_update.assert_called_with({'name': 'CPAM0002'}, {
+        '$push': {'third_party_links': {'type': "phenotips_com", 'link': 'https://phenotips.com'}}
+    },
+                                                                          return_document=True)
 
 
 def test_attach_third_party_link_analysis_does_not_exist(analysis_collection):
