### 2023-04-26 6ff9830
filtering out items that are not intended to be tested in this test
