### 2023-04-26 cf831ec
Case status in header (#71)
* first pass at this feature

* investigating warning

* fixed a warning and linting fix

* made to be a prop in the rosalution header to be in one place

* removed dead code

* added unit tests for this feature

* style fix

* lint fix

* partial change requests

* lint fix

* removed computed values

* dynamically updates the status icon when marked ready

* lint fix

* reverting back to the name of the analysis for the title in the header

this was an unintended change that I accidentally introduced.

* filtering out items that are not intended to be tested in th
[…]
