+++ frontend/test/components/RosalutionHeader.spec.js
@@ -103,6 +103,55 @@ describe('RosalutionHeaderComponent.vue', () => {
     expect(wrapper.vm.$route.path).to.equal('/rosalution/account');
   });
 
+  it('should display the proper workflow icon based on the workflow_status prop', () => {
+    [
+      {
+        workflow_status: 'Annotation',
+        expected: 'asterisk',
+      },
+      {
+        workflow_status: 'Active',
+        expected: 'book-open',
+      },
+      {
+        workflow_status: 'Approved',
+        expected: 'check',
+      },
+    ].forEach((test) => {
+      const wrapper = getMountedComponent({workflow_status: test.workflow_status});
+      const statusIcon = wrapper.get('font-awesome-icon-stub');
+      expect(statusIcon.attributes().icon).to.equal(test.expected);
+    });
+  });
+
+  it('should display the proper workflow color based on the workflow_status prop', () => {
+    [
+      {
+        workflow_status: 'Annotation',
+        expected: 'color: var(--rosalution-status-annotation);',
+      },
+      {
+        workflow_status: 'Active',
+        expected: 'color: var(--rosalution-status-active);',
+      },
+      {
+        workflow_status: 'Approved',
+        expected: 'color: var(--rosalution-status-approved);',
+      },
+    ].forEach((test) => {
+      const wrapper = getMountedComponent({workflow_status: test.workflow_status});
+      const statusIcon = wrapper.get('font-awesome-icon-stub');
+      expect(statusIcon.attributes().style).to.equal(test.expected);
+    });
+  });
+
+  it('should not render any icon when the workflow_status prop is not provided', () => {
+    const wrapper = getMountedComponent();
+    const statusIcon = wrapper.find('[data-test="status-icon"]');
+    expect(statusIcon.exists()).to.be.false;
+  });
+
+
   it('should render third party links', () => {
     const wrapper = getMountedComponent();
 
