### 2023-05-02 e9cac57
System test third party link (#73)
* added the rel attribute to enhance security of these new links

* added new tests for the third party link in header and on the cards
