### 2023-05-03 2c2934d
Resolved unit test warning and removed extra console.logs found
