### 2023-05-03 ff29bbe
resolved unit tests and removed extra console logs.
