### 2023-05-03 5525837
Updated uvicorn in backend to uvicorn[standard]==0.22.0, including the standard library for uvicorn
