### 2023-05-03 36769a1
The backend now deletes the cookie if it is invalid and the frontend end detects if the cookie has been deleted and redirects the user to the login page
