### 2023-05-03 40e94fe
Working on clearing the auth's state when logged out and unauthorized
