### 2023-05-04 aa40772
Changed logout router calls from path to name, added a logout router call for logout to clear the user state, and created a way to clear the state object
