+++ frontend/src/main.js
@@ -61,12 +61,10 @@ const router = createRouter({
 router.beforeEach(async (to) => {
   const token = authStore.getToken();
 
-  console.log("This is happening?")
-  console.log(authStore.state);
-
   if (!token && to.name !== 'login') {
     return {name: 'login'};
-  } else if (token && to.name == 'account') {
+  }
+  else if (token && to.name == 'account') {
     const response = await authStore.getAPICredentials();
 
     if (response.error) {
@@ -75,17 +73,20 @@ router.beforeEach(async (to) => {
     }
 
     authStore.saveState(response);
-  } else if (token) {
+  } 
+  else if (to.name == 'logout') {
+    authStore.clearState();
+  }
+  else if (token) {
     const response = await authStore.verifyToken();
 
     if (response.error) {
       authStore.clearState();
-      console.log(authStore.state);
       return {name: 'login'};
     }
 
     authStore.saveState(response);
-  }
+  } 
 });
 
 const app = createApp(App);
