+++ frontend/src/stores/authStore.js
@@ -9,10 +9,8 @@ const authStore = {
     clientId: '',
     clientSecret: '',
   },
-  getUser() {
-    return {...this.state};
-  },
   clearState() {
+    console.log("Clear state happening?");
     this.state.full_name = '';
     this.state.username = '';
     this.state.email = '';
@@ -29,6 +27,9 @@ const authStore = {
 
     user['client_id'] ? this.state.clientId = user['client_id'] : '';
     user['client_secret'] ? this.state.clientSecret = user['client_secret'] : '';
+  },
+  getUser() {
+    return {...this.state};
   },
   getToken() {
     if (document.cookie == '') {
