### 2023-05-05 d7ea0cd
added a test to check for status icon update in header on change
