### 2023-05-08 61c711e
Properly redirect the user to the login page if the system gets updated (#77)
* The backend now deletes the cookie if it is invalid and the frontend end detects if the cookie has been deleted and redirects the user to the login page

* Working on clearing the auth's state when logged out and unauthorized

* Changed logout router calls from path to name, added a logout router call for logout to clear the user state, and created a way to clear the state object

* Added frontend and backend tests for the security
