+++ frontend/test/views/AnalysisView.spec.js
@@ -7,9 +7,11 @@ import Analyses from '@/models/analyses.js';
 import InputDialog from '@/components/Dialogs/InputDialog.vue';
 import NotificationDialog from '@/components/Dialogs/NotificationDialog.vue';
 import SupplementalFormList from '@/components/AnalysisView/SupplementalFormList.vue';
+import SaveModal from '@/components/AnalysisView/SaveModal.vue';
 
 import inputDialog from '@/inputDialog.js';
 import notificationDialog from '@/notificationDialog.js';
+import toast from '@/toast.js';
 
 import AnalysisView from '@/views/AnalysisView.vue';
 
@@ -52,6 +54,7 @@ describe('AnalysisView', () => {
   let mockedAttachThirdPartyLink;
   let markReadyMock;
   let markActiveMock;
+  let updateAnalysisSectionsStub;
   let wrapper;
   let sandbox;
 
@@ -70,6 +73,8 @@ describe('AnalysisView', () => {
 
     markReadyMock = sandbox.stub(Analyses, 'markAnalysisReady');
     markActiveMock = sandbox.stub(Analyses, 'markAnalysisActive');
+
+    updateAnalysisSectionsStub = sandbox.stub(Analyses, 'updateAnalysisSections');
 
     wrapper = getMountedComponent();
   });
@@ -106,40 +111,99 @@ describe('AnalysisView', () => {
       );
     });
 
-    it('should mark an analysis as ready', async () => {
-      const annotatingAnalysis = fixtureData();
-      annotatingAnalysis.latest_status = 'Annotation';
-      mockedData.returns(annotatingAnalysis);
-      const wrapper = getMountedComponent();
-      await wrapper.vm.$nextTick();
-      const headerComponent = wrapper.getComponent(
-          '[data-test=analysis-view-header]',
-      );
-      const actionsProps = headerComponent.props('actions');
-      for (const action of actionsProps) {
-        if (action.text === 'Mark Ready') {
-          action.operation();
-        }
-      }
-      expect(markReadyMock.called).to.be.true;
-    });
+    describe('actions with toasts', () => {
 
-    it('should mark an analysis as active', async () => {
-      const readyAnalysis = fixtureData();
-      readyAnalysis.latest_status = 'Ready';
-      mockedData.returns(readyAnalysis);
-      wrapper = getMountedComponent();
-      await wrapper.vm.$nextTick();
-      const headerComponent = wrapper.getComponent(
-          '[data-test=analysis-view-header]',
-      );
-      const actionsProps = headerComponent.props('actions');
-      for (const action of actionsProps) {
-        if (action.text === 'Mark Active') {
-          action.operation();
+      /**
+       * Helper triggers an action based on the action text
+       * @param {VueWrapper} wrapper The Vue wrapper containing the component instance
+       * @param {string} actionText The text of the action to trigger
+       * @return {Promise} A promise that resolves when the action is triggered
+       */
+      async function triggerAction(wrapper, actionText) {
+        const headerComponent = wrapper.getComponent('[data-test=analysis-view-header]');
+        const actionsProps = headerComponent.props('actions');
+        for (const action of actionsProps) {
+          if (action.text === actionText) {
+            action.operation();
+            break;
+          }
         }
       }
-      expect(markActiveMock.called).to.be.true;
+
+      /**
+       * Helper sets up a mocked wrapper with the given latest status and returns it
+       * @param {string} latestStatus The latest status of the analysis to set
+       * @return {Promise<VueWrapper>} A promise that resolves with the mocked Vue wrapper
+       */
+      async function getMockedWrapper(latestStatus) {
+        const analysisData = fixtureData();
+        analysisData.latest_status = latestStatus;
+        mockedData.returns(analysisData);
[…]
