### 2023-05-09 8cedd50
added a test and some logic for displaying a toast for editing an analysis
caught this while writing the pull request
