### 2023-05-10 27a7f74
Sys test status in header (#78)
* modified a fixture to have an annotation card loaded in

* added a test to check for status icon update in header on change

* added a new card for seeding for testing

* whoops broke a test that needed updating

I added a new card to the seed so this needed to be updated to 6 from 5

* same as last push. this needed updating from 5 to 6

* updated analysis cards length from 5 to 6 in sys test

* removing these three properties
