+++ backend/tests/unit/repository/test_analysis_collection.py
@@ -44,7 +44,7 @@ def test_find_file_by_name(analysis_collection):
     """Tests the find_file_by_name function"""
     analysis_collection.collection.find_one.return_value = read_test_fixture("analysis-CPAM0002.json")
     actual = analysis_collection.find_file_by_name("CPAM0002", "test.txt")
-    assert actual == {'attachment_id': '633afb87fb250a6ea1569555', 'comments': 'hello world', 'filename': 'test.txt'}
+    assert actual == {'attachment_id': '633afb87fb250a6ea1569555', 'comments': 'hello world', 'name': 'test.txt'}
 
 
 def test_find_file_by_name_analysis_none(analysis_collection):
@@ -132,6 +132,12 @@ def test_attach_file_supporting_evidence(analysis_collection, cpam0002_analysis_
     assert new_evidence['type'] == 'file'
     assert 'attachment_id' in new_evidence
     assert new_evidence['attachment_id'] == 'Fake-Mongo-Object-ID-2'
+
+def test_file_exists_in_analysis(analysis_collection):
+    """Tests the file_exists_in_analysis function"""
+    analysis_collection.collection.find_one.return_value = read_test_fixture("analysis-CPAM0002.json")
+    actual = analysis_collection.file_exists_in_analysis("CPAM0002", "test.txt")
+    assert actual is True
 
 
 def test_add_pedigree_file(analysis_collection, empty_pedigree):
