+++ frontend/test/views/AnalysisView.spec.js
@@ -44,6 +44,39 @@ function getMountedComponent(props) {
   });
 }
 
+/**
+ * Helper triggers an action based on the action text
+ * @param {VueWrapper} wrapper The Vue wrapper containing the component instance
+ * @param {string} actionText The text of the action to trigger
+ * @return {Promise} A promise that resolves when the action is triggered
+ */
+async function triggerAction(wrapper, actionText) {
+  const headerComponent = wrapper.getComponent('[data-test=analysis-view-header]');
+  const actionsProps = headerComponent.props('actions');
+  for (const action of actionsProps) {
+    if (action.text === actionText) {
+      action.operation();
+      break;
+    }
+  }
+}
+
+/**
+ * Helper sets up a mocked wrapper with the given latest status and returns it
+ * @param {string} latestStatus The latest status of the analysis to set
+ * @param {object} mockedData The mocked data to be used
+ * @return {Promise<VueWrapper>} A promise that resolves with the mocked Vue wrapper
+ */
+async function getMockedWrapper(latestStatus, mockedData) {
+  const analysisData = fixtureData();
+  analysisData.latest_status = latestStatus;
+  mockedData.returns(analysisData);
+  const wrapper = getMountedComponent();
+  await wrapper.vm.$nextTick();
+  return wrapper;
+}
+
+
 describe('AnalysisView', () => {
   let mockedData;
   let pedigreeAttachMock;
@@ -54,7 +87,7 @@ describe('AnalysisView', () => {
   let mockedAttachThirdPartyLink;
   let markReadyMock;
   let markActiveMock;
-  let updateAnalysisSectionsStub;
+  let updateAnalysisSectionsMock;
   let wrapper;
   let sandbox;
 
@@ -74,7 +107,7 @@ describe('AnalysisView', () => {
     markReadyMock = sandbox.stub(Analyses, 'markAnalysisReady');
     markActiveMock = sandbox.stub(Analyses, 'markAnalysisActive');
 
-    updateAnalysisSectionsStub = sandbox.stub(Analyses, 'updateAnalysisSections');
+    updateAnalysisSectionsMock = sandbox.stub(Analyses, 'updateAnalysisSections');
 
     wrapper = getMountedComponent();
   });
@@ -111,107 +144,74 @@ describe('AnalysisView', () => {
       );
     });
 
-    describe('actions with toasts', () => {
-      /**
-       * Helper triggers an action based on the action text
-       * @param {VueWrapper} wrapper The Vue wrapper containing the component instance
-       * @param {string} actionText The text of the action to trigger
-       * @return {Promise} A promise that resolves when the action is triggered
-       */
-      async function triggerAction(wrapper, actionText) {
-        const headerComponent = wrapper.getComponent('[data-test=analysis-view-header]');
-        const actionsProps = headerComponent.props('actions');
-        for (const action of actionsProps) {
-          if (action.text === actionText) {
-            action.operation();
-            break;
-          }
-        }
-      }
+    it('should mark an analysis as ready', async () => {
+      const wrapper = await getMockedWrapper('Annotation', mockedData);
 
-      /**
-       * Helper sets up a mocked wrapper with the given latest status and returns it
-       * @param {string} latestStatus The latest status of the analysis to set
-       * @return {Promise<VueWrapper>} A promise that resolves with the mocked Vue wrapper
-       */
-      async function getMockedWrapper(latestStatus) {
-        const analysisData = fixtureData();
-        analysisData.latest_status = latestStatus;
-        mockedData.returns(analysisData);
-        const wrapper = getMountedComponent();
-        await wrapper.vm.$nextTick();
-        return wrapper;
-      }
+      await triggerAction(wrapper, 'Mark Ready');
 
[…]
