### 2023-05-15 cdee0c8
Removed file_exists_in_analysis function and refactored its uses to leverage find_file_by_name; deleted associated unit tests
