### 2023-05-16 8e64e10
Same name supporting evidence (#80)
* updated attach_supporting_evidence_file, added file_exists_in_analysis,  unit test, and had to pay some tech debt for these changes

while previously the tests were passing, in reality the functionality of finding file by name did not work.

as for the filename to name change, in our system we use name and not filename.

the change of file_id to attachment_id is the same as above. it's what we use in our system.

* running YAPF to fix formatting errors

* adding system tests for this new functionality of adding supporting evidence

* making requested chang
[…]
