### 2023-05-16 364785d
Update Annotation Configuration to include new annotations from Alliance Genome (#81)
* Now fetching the proper alliance genomic identifiers, with the correct identifiers the models are now being pulled in as expected

* Fixed tests due to more annotations in the configuration, 4 to be exact

* updated integration test annotation call values
