### 2023-05-17 f80dcb6
Reverted dataset type for HPO to text-dataset from tag-dataset; added new unit test TagDataset.spec.js; updated AnnotationView.spec.js for future tag-dataset enablement
