### 2023-05-17 465d6c8
Removed unnecessary trim operation
