### 2023-05-18 b2377ea
TagDataset for Annotation (#83)
* making progress on the pills for hpo terms component

* Renamed HpoPills.vue to TagDataset.vue and refactored to display as tags

refactored tagDataset to parse and display '; ' delimited strings as pill-shaped tags

* Reverted dataset type for HPO to text-dataset from tag-dataset; added new unit test TagDataset.spec.js; updated AnnotationView.spec.js for future tag-dataset enablement

* lint fix

* Removed unnecessary trim operation

* making the requested PR changes

* updated the style to make use of our vars

* updating the line-height as well

* lint fix
