### 2023-05-19 4bf57b3
The card is taking shape, lots of changes
