+++ frontend/src/models/analyses.js
@@ -311,6 +311,90 @@ const annotationRenderingTemporary = [
   {
     'type': 'section',
     'class': '',
+    'header': 'Chromosomal Localization',
+    'anchor': 'Chromosomal_Localization',
+    'header_datasets': [],
+    'props': {
+      'attachmentDataset': 'Chromosomal_Localization',
+      'genomicAttachmentType': 'variant',
+    },
+    'rows': [{
+      'class': '',
+      'datasets': [
+        {
+          'dataset': 'Chromosomal_Localization',
+          'genomicType': 'variant',
+          'type': 'images-dataset',
+        },
+      ],
+    }],
+  },
+  {
+    'type': 'section',
+    'class': '',
+    'header': 'Secondary Structure',
+    'anchor': 'Secondary_Structure',
+    'header_datasets': [],
+    'props': {
+      'attachmentDataset': 'Secondary Structure',
+      'genomicAttachmentType': 'variant',
+    },
+    'rows': [{
+      'class': '',
+      'datasets': [
+        {
+          'dataset': 'Secondary_Structure',
+          'genomicType': 'variant',
+          'type': 'images-dataset',
+        },
+      ],
+    }],
+  },
+  {
+    'type': 'section',
+    'class': '',
+    'header': 'Casual Variant In This Locus (ClinVar)',
+    'anchor': 'Casual_Variant',
+    'header_datasets': [],
+    'props': {
+      'attachmentDataset': 'Casual Variant In This Locus (ClinVar)',
+      'genomicAttachmentType': 'variant',
+    },
+    'rows': [{
+      'class': '',
+      'datasets': [
+        {
+          'dataset': 'Casual Variant In This Locus (ClinVar)',
+          'genomicType': 'variant',
+          'type': 'images-dataset',
+        },
+      ],
+    }],
+  },
+  {
+    'type': 'section',
+    'class': '',
+    'header': 'Variant Publications',
+    'anchor': 'Variant_Publications',
+    'header_datasets': [],
+    'props': {
+      'attachmentDataset': 'Variant Publications',
+      'genomicAttachmentType': 'variant',
+    },
+    'rows': [{
+      'class': '',
+      'datasets': [
+        {
+          'dataset': 'Variant_Publications',
+          'genomicType': 'variant',
+          'type': 'images-dataset',
+        },
+      ],
+    }],
+  },
+  {
+    'type': 'section',
+    'class': '',
     'header': 'Gene Homology/Multi-Sequence Alignment',
     'anchor': 'Gene_Homology',
     'header_datasets': [],
@@ -332,11 +416,11 @@ const annotationRenderingTemporary = [
   {
     'type': 'section',
     'class': '',
-    'header': 'Protein Expression',
-    'anchor': 'Protein_Expression',
+    'header': 'Human Gene Expression',
+    'anchor': 'Human_Gene_Expression',
     'header_datasets': [],
     'props': {
-      'attachmentDataset': 'Protein Expression',
+      'attachmentDataset': 'Human Gene Expression',
       'genomicAttachmentType': 'gene',
     },
     'rows': [{
@@ -349,6 +433,137 @@ const annotationRenderingTemporary = [
         },
       ],
     }],
+  },
+  {
+    'type': 'section',
+    'class': '',
+    'header': 'Human Gene versus Protein Expression Profile',
+    'anchor': 'Human_Gene_versus_Protein_Expression',
+    'header_datasets': [],
+    'props': {
+      'attachmentDataset': 'Human Gene versus Protein Expression Profile',
+      'genomicAttachmentType': 'gene',
+    },
+    'rows': [{
+      'class': '',
+      'datasets': [
+        {
[…]
