### 2023-05-22 bd34343
updating the sidebar style to remained a fixed size, updated the logic for removing '_' from the anchors when displaying them
