+++ system-tests/e2e/case_annotation_section_anchors.cy.js
@@ -8,7 +8,7 @@ describe('case_annotation_display_transcripts.cy.js', () => {
 
   it('navigates to the annotation sections via anchor links', () => {
     const anchorLinks = ['Gene', 'Variant', 'Chromosomal Localization', 'Secondary Structure', 'Casual Variant',
-      'Variant Publications', 'Gene Homology', 'Human Gene_Expression', 'Human Gene_versus_Protein_Expression',
+      'Variant Publications', 'Gene Homology', 'Human Gene Expression', 'Human Gene versus Protein Expression',
       'Expression Profiles', 'Orthology', 'Mouse Model', 'Rat Model', 'Zebrafish Model', 'Worm Model', 'Modelability',
       'Druggability'];
     const expectedSidebarLinks = [...anchorLinks];
