### 2023-05-24 f7432c1
PHI warning, added warning option for inputDialog
