### 2023-05-24 a3c4b84
made requested PR changes, updating variant to hgvs_variant for attachment/genomic Type type, update styling to use variable set in main.css
