### 2023-05-24 73924e3
reduced the CSS footprint for the gene box component and the dom to be simpler
