### 2023-05-25 d11e48a
Updated the animal model fixtures to be correct, still working on the card design as new data comes in
