### 2023-05-25 74a5554
Update frontend/src/models/analyses.js
Co-authored-by: Angelina Uno-Antonison <ange.unoantonison@gmail.com>
Signed-off-by: Alex Moss <kuthedk@users.noreply.github.com>
