### 2023-05-25 6efaa29
Update case_annotation_section_anchors.cy.js
