+++ frontend/src/models/analyses.js
@@ -311,6 +311,90 @@ const annotationRenderingTemporary = [
   {
     'type': 'section',
     'class': '',
+    'header': 'Chromosomal Localization',
+    'anchor': 'Chromosomal_Localization',
+    'header_datasets': [],
+    'props': {
+      'attachmentDataset': 'Chromosomal_Localization',
+      'genomicAttachmentType': 'hgvs_variant',
+    },
+    'rows': [{
+      'class': '',
+      'datasets': [
+        {
+          'dataset': 'Chromosomal_Localization',
+          'genomicType': 'hgvs_variant',
+          'type': 'images-dataset',
+        },
+      ],
+    }],
+  },
+  {
+    'type': 'section',
+    'class': '',
+    'header': 'Secondary Structure',
+    'anchor': 'Secondary_Structure',
+    'header_datasets': [],
+    'props': {
+      'attachmentDataset': 'Secondary_Structure',
+      'genomicAttachmentType': 'hgvs_variant',
+    },
+    'rows': [{
+      'class': '',
+      'datasets': [
+        {
+          'dataset': 'Secondary_Structure',
+          'genomicType': 'hgvs_variant',
+          'type': 'images-dataset',
+        },
+      ],
+    }],
+  },
+  {
+    'type': 'section',
+    'class': '',
+    'header': 'Causal Variant In This Locus (ClinVar)',
+    'anchor': 'Causal_Variant',
+    'header_datasets': [],
+    'props': {
+      'attachmentDataset': 'Causal_Variant_In_This_Locus_ClinVar',
+      'genomicAttachmentType': 'hgvs_variant',
+    },
+    'rows': [{
+      'class': '',
+      'datasets': [
+        {
+          'dataset': 'Causal_Variant_In_This_Locus_ClinVar',
+          'genomicType': 'hgvs_variant',
+          'type': 'images-dataset',
+        },
+      ],
+    }],
+  },
+  {
+    'type': 'section',
+    'class': '',
+    'header': 'Variant Publications',
+    'anchor': 'Variant_Publications',
+    'header_datasets': [],
+    'props': {
+      'attachmentDataset': 'Variant_Publications',
+      'genomicAttachmentType': 'hgvs_variant',
+    },
+    'rows': [{
+      'class': '',
+      'datasets': [
+        {
+          'dataset': 'Variant_Publications',
+          'genomicType': 'hgvs_variant',
+          'type': 'images-dataset',
+        },
+      ],
+    }],
+  },
+  {
+    'type': 'section',
+    'class': '',
     'header': 'Gene Homology/Multi-Sequence Alignment',
     'anchor': 'Gene_Homology',
     'header_datasets': [],
@@ -332,23 +416,170 @@ const annotationRenderingTemporary = [
   {
     'type': 'section',
     'class': '',
-    'header': 'Protein Expression',
-    'anchor': 'Protein_Expression',
+    'header': 'Human Gene Expression',
+    'anchor': 'Human_Gene_Expression',
     'header_datasets': [],
     'props': {
-      'attachmentDataset': 'Protein Expression',
+      'attachmentDataset': 'Protein_Expression',
       'genomicAttachmentType': 'gene',
     },
     'rows': [{
       'class': '',
       'datasets': [
         {
-          'dataset': 'Protein Expression',
+          'dataset': 'Protein_Expression',
           'genomicType': 'gene',
           'type': 'images-dataset',
         },
       ],
     }],
+  },
+  {
+    'type': 'section',
+    'class': '',
+    'header': 'Human Gene versus Protein Expression Profile',
+    'anchor': 'Human_Gene_versus_Protein_Expression',
+    'header_datasets': [],
+    'props': {
+      'attachmentDataset': 'Human_Gene_versus_Protein_Expression_Profile',
[…]
