### 2023-05-30 ad1ebd2
Case input phi warning (#85)
* PHI warning, added warning option for inputDialog

* larger drag & drop box

* locking the modal with if there is a warningText set for the dialog

* unit tests

* making some requested changes

* changing from warningTest to message
