+++ frontend/test/components/AnnotationView/AllianceGenomeCard.spec.js
@@ -5,6 +5,11 @@ import AllianceGenomeCard from '@/components/AnnotationView/AllianceGenomeCard.v
 
 import {FontAwesomeIcon} from '@fortawesome/vue-fontawesome';
 
+/**
+ * helper function that shallow mounts and returns the rendered component
+ * @param {propsData} propsData props for testing to overwrite default props
+ * @return {VueWrapper} returns a shallow mounted using props
+ */
 function getMountedComponent(propsData) {
   const defaultPropsData = {
     model: fakeModelProp,
@@ -35,11 +40,79 @@ describe('AllianceGenomeCard.vue', () => {
     expect(cardBackground.html()).toContain('[background:] involves: 129S4/SvJae * C57BL/6');
   });
 
-  it('Displays the Experimental Condition section with black text because it has a property', () => {
+  it('Displays the Experimental Condition section with black text because it has list items', () => {
     const wrapper = getMountedComponent();
     const experimentalSection = wrapper.find('[data-test=model-section-condition]');
 
     expect(experimentalSection.attributes().style).toContain('color: black;');
+  });
+
+  it('Shows a list of conditions under the experimental conditions section', () => {
+    const wrapper = getMountedComponent();
+    const experimentalSectionList = wrapper.findAll('[data-test=model-list-condition]');
+
+    expect(experimentalSectionList.length).to.equal(1);
+  });
+
+  it('Shows the Associated Human Diseases as gray because it has list items', () => {
+    const wrapper = getMountedComponent();
+    const diseasesSection = wrapper.find('[data-test=model-section-disease]');
+
+    expect(diseasesSection.attributes().style).toContain('color: black;');
+  });
+
+  it('Shows the Associated Human Diseases as gray because there are no list items', () => {
+    const wrapper = getMountedComponent({model: fakeEmptyModelProp});
+    const diseasesSection = wrapper.find('[data-test=model-section-disease]');
+
+    expect(diseasesSection.attributes().style).toContain('--rosalution-grey-300');
+  });
+
+  it('Does not display a list as there are no diseases to show', () => {
+    const wrapper = getMountedComponent({model: fakeEmptyModelProp});
+    const diseasesList = wrapper.find('[data-test=model-list-disease]');
+
+    expect(diseasesList.exists()).to.be.false;
+  });
+
+  it('Displays the Associated Phenotypes section with black text because it has list items', () => {
+    const wrapper = getMountedComponent();
+    const phenotypesSection = wrapper.find('[data-test=model-section-phenotype]');
+
+    expect(phenotypesSection.attributes().style).toContain('color: black;');
+  });
+
+  it('Shows the Associated Phenotypes section as gray because there are no items', () => {
+    const wrapper = getMountedComponent({model: fakeEmptyModelProp});
+    const phenotypesSection = wrapper.find('[data-test=model-section-phenotype]');
+
+    expect(phenotypesSection.attributes().style).toContain('--rosalution-grey-300');
+  });
+
+  it('Displays Phenotype lists with frequent term sub-headers ', () => {
+    const wrapper = getMountedComponent({model: fakeModelProp});
+    const phenotypesList = wrapper.findAll('[data-test=model-list-phenotype]');
+
+    const firstTerm = phenotypesList[0].find('span');
[…]
