### 2023-06-04 uab-cgds-worthey/rosalution#87 @suragnair
State of the field section missing in paper
According to the JOSS review guidelines, the authors should provide a section titled 'State of the field' which describes how this software compares to other commonly-used packages. This section is currently missing from the [paper](https://github.com/openjournals/joss-reviews/issues/5443).
