### 2023-06-04 #5443 @suragnair
@SeriousHorncat would it be possible to include either a link to a demo website (pre-populated with a few examples) or a video covering basic usage in the GitHub README? This would allow potential users to get a quick overview of the functionality and how to use Rosalution.
