### 2023-06-05 666de30
Updated the AllianceGenomeCard.spec.js toContain to be to.include
