### 2023-06-05 b8adf95
Create Animal Model Card to represent the animal models from Alliance Genome (#86)
* Creating the files to house alliance genome card data

* The card is taking shape, lots of changes

* Playing with making color and icons for the model card list

* Updated the animal model fixtures to be correct, still working on the card design as new data comes in

* Adding scrolling to the cards

* Changed the model annotations to get a limit of 100 models, rather than 20 models. Updated the fixtures to contain ALL models

* experimental conditions is working

* Started on the AllianceGenomeCard tests and 
[…]
