+++ system-tests/e2e/case_annotation_section_anchors.cy.js
@@ -9,8 +9,8 @@ describe('case_annotation_display_transcripts.cy.js', () => {
   it('navigates to the annotation sections via anchor links', () => {
     const anchorLinks = ['Gene', 'Variant', 'Chromosomal Localization', 'Secondary Structure', 'Causal Variant',
       'Variant Publications', 'Gene Homology', 'Human Gene Expression', 'Human Gene versus Protein Expression',
-      'Expression Profiles', 'Orthology', 'Mouse_Animal_Models', 'Rat_Animal_Models', 'Zebrafish_Animal_Models', 'C_Elegens_Animal_Models', 'Modelability',
-      'Druggability'];
+      'Expression Profiles', 'Orthology', 'Rattus norvegicus Model System', 'Mus musculus Model System',
+      'Danio rerio Model System', 'C Elegens Model System', 'Modelability', 'Druggability'];
     const expectedSidebarLinks = [...anchorLinks];
 
     cy.get('.sidebar').find('a').each(($el) => {
