### 2023-06-06 642187d
resolved linting errors in frontend that occured during the merge
