+++ frontend/test/views/AnnotationView.spec.js
@@ -103,7 +103,6 @@ describe('AnnotationView', () => {
   it('renders expected number of sections ', () => {
     const sections = wrapper.findAllComponents(AnnotationSection);
     expect(sections.length).to.equal(17);
-
   });
 
   it('renders each section with an Id to be used for an achor', () => {
