+++ frontend/test/components/AnnotationView/AllianceGenomeCard.spec.js
@@ -1,0 +1,192 @@
+import {describe, expect, it} from 'vitest';
+import {shallowMount} from '@vue/test-utils';
+
+import AllianceGenomeCard from '@/components/AnnotationView/AllianceGenomeCard.vue';
+
+import {FontAwesomeIcon} from '@fortawesome/vue-fontawesome';
+
+/**
+ * helper function that shallow mounts and returns the rendered component
+ * @param {propsData} propsData props for testing to overwrite default props
+ * @return {VueWrapper} returns a shallow mounted using props
+ */
+function getMountedComponent(propsData) {
+  const defaultPropsData = {
+    model: fakeModelProp,
+  };
+
+  return shallowMount(AllianceGenomeCard, {
+    props: {...defaultPropsData, ...propsData},
+    global: {
+      components: {
+        'font-awesome-icon': FontAwesomeIcon,
+      },
+    },
+  });
+}
+
+describe('AllianceGenomeCard.vue', () => {
+  it('Displays a card with the proper title', () => {
+    const wrapper = getMountedComponent();
+    const cardHeader = wrapper.find('[data-test=model-header]');
+
+    expect(cardHeader.html()).to.include('lmna<sup>bw25/bw25</sup>');
+  });
+
+  it('Displays a card with a proper background', () => {
+    const wrapper = getMountedComponent();
+    const cardBackground = wrapper.find('[data-test=model-background]');
+
+    expect(cardBackground.html()).to.include('[background:] involves: 129S4/SvJae * C57BL/6');
+  });
+
+  it('Displays the Experimental Condition section with black text because it has list items', () => {
+    const wrapper = getMountedComponent();
+    const experimentalSection = wrapper.find('[data-test=model-section-condition]');
+
+    expect(experimentalSection.attributes().style).to.include('color: black;');
+  });
+
+  it('Shows a list of conditions under the experimental conditions section', () => {
+    const wrapper = getMountedComponent();
+    const experimentalSectionList = wrapper.findAll('[data-test=model-list-condition]');
+
+    expect(experimentalSectionList.length).to.equal(1);
+  });
+
+  it('Shows the Associated Human Diseases as gray because it has list items', () => {
+    const wrapper = getMountedComponent();
+    const diseasesSection = wrapper.find('[data-test=model-section-disease]');
+
+    expect(diseasesSection.attributes().style).to.include('color: black;');
+  });
+
+  it('Shows the Associated Human Diseases as gray because there are no list items', () => {
+    const wrapper = getMountedComponent({model: fakeEmptyModelProp});
+    const diseasesSection = wrapper.find('[data-test=model-section-disease]');
+
+    expect(diseasesSection.attributes().style).to.include('--rosalution-grey-300');
+  });
+
+  it('Does not display a list as there are no diseases to show', () => {
+    const wrapper = getMountedComponent({model: fakeEmptyModelProp});
+    const diseasesList = wrapper.find('[data-test=model-list-disease]');
+
+    expect(diseasesList.exists()).to.be.false;
+  });
+
+  it('Displays the Associated Phenotypes section with black text because it has list items', () => {
+    const wrapper = getMountedComponent();
+    const phenotypesSection = wrapper.find('[data-test=model-section-phenotype]');
+
+    expect(phenotypesSection.attributes().style).to.include('color: black;');
+  });
+
+  it('Shows the Associated Phenotypes section as gray because there are no items', () => {
[…]
