+++ frontend/src/main.js
@@ -23,7 +23,7 @@ import {
   faAsterisk, faPause, faCheck, faX, faUser, faUsers, faUserGroup, faCalendar, faBookOpen, faList, faLayerGroup,
   faBoxArchive, faQuestion, faClock, faClipboardCheck, faMagnifyingGlass, faChevronDown, faChevronRight,
   faUpRightFromSquare, faCirclePlus, faPencil, faEllipsisVertical, faLink, faXmark, faUserDoctor, faPaperclip, faPlus,
-  faAnglesRight, faFileImage, faFileCirclePlus, faMountainSun,
+  faAnglesRight, faFileImage, faFileCirclePlus, faMountainSun, faArrowUp, faArrowDown,
 } from '@fortawesome/free-solid-svg-icons';
 import {
   faCopy, faFile, faComment, faCircleCheck, faCircleQuestion, faCircleXmark, faImages,
@@ -34,7 +34,8 @@ library.add(
     faLayerGroup, faBoxArchive, faQuestion, faClock, faClipboardCheck, faChevronDown, faChevronRight,
     faUpRightFromSquare, faCopy, faCirclePlus, faFile, faComment, faPencil, faEllipsisVertical, faLink, faXmark,
     faUserDoctor, faPaperclip, faPlus, faAnglesRight, faCircleCheck, faCircleQuestion, faCircleXmark, faImages,
-    faFileImage, faFileCirclePlus, faMountainSun);
+    faFileImage, faFileCirclePlus, faMountainSun, faArrowUp, faArrowDown,
+);
 
 // The NotFoundView should always be last because it's an ordered array.
 /* Placeholder route, name subject to change */
