### 2023-06-06 be369a8
Alliance genome models into new configuration (#88)
* Update Annotation Configuration to include new annotations from Alliance Genome (#81)
* Now fetching the proper alliance genomic identifiers, with the correct identifiers the models are now being pulled in as expected
* Create Animal Model Card to represent the animal models from Alliance Genome (#86)
* Adding scrolling to the cards
* Changed the model annotations to get a limit of 100 models, rather than 20 models. Updated the fixtures to contain ALL models
* Removed display: block, position: relative from card-base. Changed li margin-left
[…]
