### 2023-06-07 c65480c
Migration script saving changes in database. Debugging comments removed
