### 2023-06-07 35a50eb
Update etc/fixtures/migrations/migrate-additional-sections.js
Added suggestion to update section name to match configuration

Co-authored-by: Angelina Uno-Antonison <ange.unoantonison@gmail.com>
Signed-off-by: Rabab Fatima <fr.rabab@gmail.com>
