### 2023-06-08 12c9465
Updated annotation configuration to have both automated and specificed alliance genome animal summaries, updated the fixtures to include both summaries and are accurate to the gene identifier, and updated the annotation rendering configuration to display the new annotations
