+++ backend/src/routers/analysis_router.py
@@ -4,7 +4,7 @@ import json
 
 from typing import List, Optional, Union
 
-from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, File, UploadFile, Form, Security
+from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, File, status, UploadFile, Form, Response, Security
 from fastapi.responses import StreamingResponse
 
 from ..core.annotation import AnnotationService
@@ -55,32 +55,6 @@ def get_genomic_units(analysis_name: str, repositories=Depends(database)):
         return repositories["analysis"].get_genomic_units(analysis_name)
     except ValueError as exception:
         raise HTTPException(status_code=404, detail=str(exception)) from exception
-
-
-@router.post("/import", response_model=Analysis)
-async def import_phenotips_json(
-    background_tasks: BackgroundTasks,
-    phenotips_input: BasePhenotips,
-    repositories=Depends(database),
-    annotation_task_queue=Depends(annotation_queue),
-    username: VerifyUser = Security(get_current_user)
-):
-    """Imports the phenotips.json file into the database"""
-    phenotips_importer = PhenotipsImporter(repositories["analysis"], repositories["genomic_unit"])
-    try:
-        new_analysis = phenotips_importer.import_phenotips_json(phenotips_input)
-        new_analysis['timeline'].append(Event.timestamp_create_event(username).dict())
-        repositories['analysis'].create_analysis(new_analysis)
-
-    except ValueError as exception:
-        raise HTTPException(status_code=409) from exception
-
-    analysis = Analysis(**new_analysis)
-    annotation_service = AnnotationService(repositories["annotation_config"])
-    annotation_service.queue_annotation_tasks(analysis, annotation_task_queue)
-    background_tasks.add_task(AnnotationService.process_tasks, annotation_task_queue, repositories['genomic_unit'])
-
-    return new_analysis
 
 
 @router.post("/import_file", response_model=Analysis)
@@ -152,21 +126,41 @@ def download(analysis_name: str, file_name: str, repositories=Depends(database))
     return StreamingResponse(repositories['bucket'].stream_analysis_file_by_id(file['attachment_id']))
 
 
-@router.post("/{analysis_name}/attach/section/image", response_model=Section, response_model_exclude_none=True)
-def upload_pedigree(analysis_name: str, upload_file: UploadFile = File(...), section_name: str = Form(...), field_name: str = Form(...), repositories=Depends(database)):
-    """ Specifically accepts a file to save a pedigree image file to mongo """
-    new_file_object_id = repositories["bucket"].save_file(
-        upload_file.file, upload_file.filename, upload_file.content_type
-    )
+@router.post("/{analysis_name}/section/attach/image/", response_model=Section, response_model_exclude_none=True)
+def upload_section_image(
+    response: Response,
+    analysis_name: str,
+    upload_file: UploadFile = File(...),
+    section_name: str = Form(...),
+    field_name: str = Form(...),
+    repositories=Depends(database)
+):
+    """ Saves the uploaded image it to the specified field_name in the analysis's section."""
+    
+    try:
+      new_file_object_id = repositories["bucket"].save_file(
+          upload_file.file, upload_file.filename, upload_file.content_type
+      )
+    except Exception as exception:
+        raise HTTPException(status_code=500, detail=str(exception)) from exception
 
     updated_section = repositories["analysis"].add_section_image(analysis_name, section_name, field_name, new_file_object_id)
+    
+    response.status_code = status.HTTP_201_CREATED
+    
     return updated_section
 
 
-@router.put("/{analysis_name}/update/pedigree", response_model=Section)
[…]
