### 2023-06-14 e92cd25
Investigate why CPAM0084 breaks when re-running annotation from the api docs (#91)
* Updated the fixtures for CPAM0084

* Updated python tests
