+++ backend/tests/unit/repository/test_annotation_collection.py
@@ -8,13 +8,13 @@ def test_get_datasets_configuration_by_type(annotation_collection):
     """Tests getting the datasets for the provided types of genomic units"""
     types = set({GenomicUnitType.GENE, GenomicUnitType.HGVS_VARIANT})
     datasets = annotation_collection.datasets_to_annotate_by_type(types)
-    assert len(datasets) == 33
+    assert len(datasets) == 37
 
 
 def test_get_datasets_to_annotate_for_units(annotation_collection, genomic_units_for_annotation):
     """Tests if the configuration for datasets is return as expected"""
     actual_configuration = annotation_collection.datasets_to_annotate_for_units(genomic_units_for_annotation)
-    assert len(actual_configuration["gene"]) == 23
+    assert len(actual_configuration["gene"]) == 27
     assert len(actual_configuration["hgvs_variant"]) == 10
 
 
