+++ backend/tests/unit/core/test_annotate.py
@@ -11,7 +11,7 @@ def test_queuing_annotations_for_genomic_units(cpam0046_analysis, annotation_col
     annotation_service = AnnotationService(annotation_collection)
     mock_queue = Mock()
     annotation_service.queue_annotation_tasks(cpam0046_analysis, mock_queue)
-    assert mock_queue.put.call_count == 33
+    assert mock_queue.put.call_count == 37
 
 
 # Patching the temporary helper method that is writing to a file, this will be
@@ -52,11 +52,11 @@ def test_processing_cpam0046_annotation_tasks(
     AnnotationService.process_tasks(cpam0046_annotation_queue, mock_genomic_unit_collection)
     assert cpam0046_annotation_queue.empty()
 
-    assert http_task_annotate.call_count == 27
+    assert http_task_annotate.call_count == 31
     assert none_task_annotate.call_count == 0
     assert forge_task_annotate.call_count == 6
 
-    assert annotate_extract_mock.call_count == 33
+    assert annotate_extract_mock.call_count == 37
 
 
 @patch("src.core.annotation.log_to_file")
@@ -86,11 +86,11 @@ def test_processing_cpam0002_annotations_tasks(
 
     AnnotationService.process_tasks(cpam0002_annotation_queue, mock_genomic_unit_collection)
 
-    assert http_task_annotate.call_count == 27
+    assert http_task_annotate.call_count == 31
     assert forge_task_annotate.call_count == 6
     assert none_task_annotate.call_count == 0
 
-    assert annotate_extract_mock.call_count == 33
+    assert annotate_extract_mock.call_count == 37
 
     mock_genomic_unit_collection.annotate_genomic_unit.assert_called()
 
