+++ frontend/src/models/analyses.js
@@ -506,10 +506,17 @@ const annotationRenderingTemporary = [
       'class': '',
       'datasets': [
         {
-          'dataset': 'Model Systems - Rat',
+          'dataset': 'Rat_Alliance_Genome_Automated_Summary',
           'type': 'text-dataset',
           'props': {
-            'label': 'Rat',
+            'label': 'AG Summary',
+          },
+        },
+        {
+          'dataset': 'Rat_Alliance_Genome_RGD_Summary',
+          'type': 'text-dataset',
+          'props': {
+            'label': 'RGD Summary',
           },
         },
         {
@@ -529,10 +536,17 @@ const annotationRenderingTemporary = [
       'class': '',
       'datasets': [
         {
-          'dataset': 'Model Systems - Mouse - Automated',
+          'dataset': 'Mouse_Alliance_Genome_Automated_Summary',
           'type': 'text-dataset',
           'props': {
-            'label': 'Mouse - Automated',
+            'label': 'AG Summary',
+          },
+        },
+        {
+          'dataset': 'Mouse_Alliance_Genome_MGI_Summary',
+          'type': 'text-dataset',
+          'props': {
+            'label': 'MGI Summary',
           },
         },
         {
@@ -552,10 +566,17 @@ const annotationRenderingTemporary = [
       'class': '',
       'datasets': [
         {
-          'dataset': 'Model Systems - Zebrafish - Automated',
+          'dataset': 'Zebrafish_Alliance_Genome_Automated_Summary',
           'type': 'text-dataset',
           'props': {
-            'label': 'Zebrafish - Automated',
+            'label': 'AG Summary',
+          },
+        },
+        {
+          'dataset': 'Zebrafish_Alliance_Genome_ZFIN_Summary',
+          'type': 'text-dataset',
+          'props': {
+            'label': 'ZFIN Summary',
           },
         },
         {
@@ -575,10 +596,17 @@ const annotationRenderingTemporary = [
       'class': '',
       'datasets': [
         {
-          'dataset': 'Model Systems - C-Elegens - Automated',
+          'dataset': 'C-Elegens_Alliance_Genome_Automated_Summary',
           'type': 'text-dataset',
           'props': {
-            'label': 'C-Elegens - Automated',
+            'label': 'AG Summary',
+          },
+        },
+        {
+          'dataset': 'C-Elegens_Alliance_Genome_WB_Summary',
+          'type': 'text-dataset',
+          'props': {
+            'label': 'WB Summary',
           },
         },
         {
