### 2023-06-15 d30e159
As a Computational Biologist, I want to see the various summaries for animal models in the relevant section so I can have all the information consolidated in one location (#92)
* Updated annotation configuration to have both automated and specificed alliance genome animal summaries, updated the fixtures to include both summaries and are accurate to the gene identifier, and updated the annotation rendering configuration to display the new annotations

* Updated the python tests to include the 4 extra annotations in the configuration

* updated the existing tests and removed the mongo ids from t
[…]
