### 2023-06-16 41744ba
Added new sections to analysis
