### 2023-06-16 bcc15e5
Refactoring to make the emits work, starting work on the updating section image in the backend
