### 2023-06-16 b15667d
tests for the status filter feature
