### 2023-06-20 a6fe80c
Replaced the component back into the sectionbox and am chaining emits from sectionimage to sectionbox to analysis view
