### 2023-06-21 96f8513
Refactored event name from 'filtered-statuses' to 'filtered-changed' and updated handler to 'filteredUpdated' for better clarity
