+++ .github/workflows/common-static-analysis.yml
@@ -4,7 +4,7 @@ on:
   pull_request:
     types:
       - opened
-      # - synchronize
+      - synchronize
       - reopened
       - ready_for_review
       - converted_to_draft
