### 2023-06-22 a6e1a8a
Experimenting with a conditional to prevent running jobs multiple times on push and pull requests yet still run for PR's for forks.
