+++ .github/workflows/common-static-analysis.yml
@@ -19,6 +19,7 @@ env:
 jobs:
   docker-shell-markdown-static-analysis:
     runs-on: ubuntu-latest
+    if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.event.pull_request.base.repo.full_name
       
     steps:
     - uses: actions/checkout@v3
