+++ .github/workflows/python.yml
@@ -11,6 +11,8 @@ jobs:
   python-ci:
 
     runs-on: ubuntu-latest
+    if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.event.pull_request.base.repo.full_name
+
     strategy:
       fail-fast: false
       matrix:
