+++ .github/workflows/nodejs.yml
@@ -10,6 +10,7 @@ jobs:
   nodejs-ci:
 
     runs-on: ubuntu-latest
+    if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.event.pull_request.base.repo.full_name
 
     strategy:
       matrix:
