### 2023-06-22 9ca4fce
Experimenting with the double trigger.
