+++ README.md
@@ -4,6 +4,8 @@
     <img src="rosalution_title.svg" alt="Rosalution Logo and Title" width="400"/>
 </p>
 
+Is it preventing the double trigger?
+
 ![license-badge](https://badgen.net/badge/license/GPLv3/blue)
 [![status](https://joss.theoj.org/papers/8f7091da9fa801c97f60a4af368c7b3a/status.svg)](https://joss.theoj.org/papers/8f7091da9fa801c97f60a4af368c7b3a)
 ![example](https://github.com/uab-cgds-worthey/rosalution/actions/workflows/python.yml/badge.svg)
