### 2023-06-22 b660ca3
Tagging the version of ubuntu being used for ubuntu-latest for consistentcy. ubuntu-22.04 is the current long term support (LTS) version.
