+++ .github/workflows/common-static-analysis.yml
@@ -18,7 +18,7 @@ env:
 
 jobs:
   docker-shell-markdown-static-analysis:
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-22.04
     if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.event.pull_request.base.repo.full_name
       
     steps:
