### 2023-06-23 116fea8
Updated the analysis image system tests, changed system-tests/e2e/attach_pedigree_image.cy.js to be attach_analysis_section_image.cy.js
