### 2023-06-26 367e1d2
Update common-static-analysis.yml
Signed-off-by: Angelina Uno-Antonison <ange.unoantonison@gmail.com>
