+++ .github/workflows/common-static-analysis.yml
@@ -1,14 +1,5 @@
 name: Markdown, Docker, and Shell Static Analysis
-on:
-  push:
-  pull_request:
-    types:
-      - opened
-      - synchronize
-      - reopened
-      - ready_for_review
-      - converted_to_draft
-  workflow_dispatch:
+on: [push, pull_request, workflow_dispatch]
 
 env:
   MARKDOWNLINT_CONFIG: etc/static-analysis/markdownlint.json
