### 2023-06-26 b745bf6
Apply suggestions from code review
Co-authored-by: James Scherer <jscherer@uab.edu>
Signed-off-by: Angelina Uno-Antonison <angelina.antonison@gmail.com>
