### 2023-06-26 c5d8dbc
Update README.md
Removed line causing the linting to break.

Signed-off-by: Angelina Uno-Antonison <angelina.antonison@gmail.com>
