### 2023-06-26 b431547
Experimenting with the double trigger. (#97)
* Experimenting with the double trigger.

* Tagging the version of ubuntu being used for ubuntu-latest for consistentcy.
ubuntu-22.04 is the current long term support (LTS) version.

* Update common-static-analysis.yml

Signed-off-by: Angelina Uno-Antonison <ange.unoantonison@gmail.com>

* Apply suggestions from code review

Co-authored-by: James Scherer <jscherer@uab.edu>
Signed-off-by: Angelina Uno-Antonison <angelina.antonison@gmail.com>

* Update README.md

Removed line causing the linting to break.

Signed-off-by: Angelina Uno-Antonison <angeli
[…]
