+++ .github/workflows/common-static-analysis.yml
@@ -1,14 +1,5 @@
 name: Markdown, Docker, and Shell Static Analysis
-on:
-  push:
-  pull_request:
-    types:
-      - opened
-      - synchronize
-      - reopened
-      - ready_for_review
-      - converted_to_draft
-  workflow_dispatch:
+on: [push, pull_request, workflow_dispatch]
 
 env:
   MARKDOWNLINT_CONFIG: etc/static-analysis/markdownlint.json
@@ -18,7 +9,7 @@ env:
 
 jobs:
   docker-shell-markdown-static-analysis:
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-22.04
     if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.event.pull_request.base.repo.full_name
       
     steps:
