### 2023-06-26 90e17d6
PR-CI/CD experiment status (#98)
* Matching quac system test

* Re-enable synchronize

* Experimenting with a conditional to prevent running jobs multiple times
on push and pull requests yet still run for PR's for forks.

* Experimenting with the double trigger. (#97)

* Experimenting with the double trigger.

* Tagging the version of ubuntu being used for ubuntu-latest for consistentcy.
ubuntu-22.04 is the current long term support (LTS) version.

* Update common-static-analysis.yml

Signed-off-by: Angelina Uno-Antonison <ange.unoantonison@gmail.com>

* Apply suggestions from code review

Co-
[…]
