+++ .github/workflows/nodejs.yml
@@ -9,7 +9,8 @@ defaults:
 jobs:
   nodejs-ci:
 
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-22.04
+    if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.event.pull_request.base.repo.full_name
 
     strategy:
       matrix:
