+++ .github/workflows/python.yml
@@ -10,7 +10,9 @@ env:
 jobs:
   python-ci:
 
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-22.04
+    if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.event.pull_request.base.repo.full_name
+
     strategy:
       fail-fast: false
       matrix:
