+++ .github/workflows/system-tests.yml
@@ -4,7 +4,8 @@ on: [push, pull_request, workflow_dispatch]
 
 jobs:
   system-tests:
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-22.04
+    if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.event.pull_request.base.repo.full_name
     
     steps:
     - uses: actions/checkout@v3
