### 2023-06-26 6f1f942
Knocked out some system tests, working on the others. I think there's a bug, need to check
