### 2023-06-27 01e2ad1
Adding the remove annotation and annotation queueing scripts
