### 2023-06-27 ae1c380
Updated the analysis annotation script to give relevant output
