### 2023-06-27 49f769f
fixed collasping of the section boxes in annotationview and made preliminary changes to the system test for collapsing
