+++ system-tests/e2e/view_collapsible_case_information.cy.js
@@ -6,17 +6,21 @@ describe('view_collapsible_case_information.cy.js', () => {
   });
 
   it('can view the case information when expanded and not when its collapsed', () => {
+    // TODO: Come back and fix after the blank line items are visible.
     cy.get('#Brief > div > [data-test="Goals"] > .section-content').should('be.visible');
     cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content').should('be.visible');
     cy.get('#Clinical_History > div > [data-test="Testing"] > .section-content').should('be.visible');
     cy.get(':nth-child(10) > .values > :nth-child(3)').should('be.visible');
-    cy.get('#Brief > tbody > .section-header > .section-header-content > .collapsable-logo > .svg-inline--fa').click();
+    
+    cy.get('#Brief > div > .rosalution-section-header > .rosalution-header-right-icons > .collapsable-logo > .svg-inline--fa').click();
     cy.get(
-        '#Clinical_History > tbody > .section-header > .section-header-content > .collapsable-logo > .svg-inline--fa')
+        '#Clinical_History > div > .rosalution-section-header > .rosalution-header-right-icons > .collapsable-logo > .svg-inline--fa')
         .click();
-    cy.get('#Brief > tbody > :nth-child(4) > :nth-child(1) > .field').should('not.be.visible');
-    cy.get('#Brief > tbody > :nth-child(5) > .values > [data-test="value-row"]').should('not.be.visible');
-    cy.get('#Clinical_History > tbody > :nth-child(4) > :nth-child(1) > .field').should('not.be.visible');
-    cy.get(':nth-child(10) > .values > :nth-child(3)').should('not.be.visible');
+
+    
+    cy.get('#Brief > div > [data-test="Goals"] > .section-content').should('not.be.visible');
+    cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content').should('not.be.visible');
+    
+    cy.get('#Clinical_History > div > [data-test="Testing"] > .section-content').should('not.be.visible');
   });
 });
