### 2023-06-28 c68e900
Updated the integration tests for backend
