### 2023-06-29 c7bc1cb
Fixed most of the linting for python
