### 2023-06-29 c7c5b80
Refactored to not require the force re-render for frontend images
