### 2023-06-30 9bb9749
As a Developer, provide a way to remove all automatic annotations from analyses, so they can be re-annotated with the correct information (#99)
* Adding the remove annotation and annotation queueing scripts

* Adding documentation and run feedback

* Updated the analysis annotation script to give relevant output

* Fixed the annotation config that broke annotation

* Linting should be passing

* Removed debugging print statements

* Added a note to the annotate-all-existing-analyses.sh that it may need to be run more than once

* Removed a space
