+++ system-tests/e2e/edit_case_analysis.cy.js
@@ -15,16 +15,20 @@ describe('edit_case_analysis.cy.js', () => {
   // });
 
   it('edits a value & saves to the analysis', () => {
-    cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content > [data-test="editable-value"]').click();
-    cy.get('#Brief > tbody > :nth-child(5) > .values > [data-test="editable-value"]').type('test');
-    cy.get('#Brief > tbody > :nth-child(5) > .values > [data-test="editable-value"]').should('contain', 'test');
+    cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content > [data-test="editable-value"]')
+      .click();
+    cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content > [data-test="editable-value"]')
+      .type('test');
+    cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content > [data-test="editable-value"]')
+      .should('contain', 'test');
     cy.get('[data-test="save-edit-button"]').should('exist');
     cy.get('[data-test="save-edit-button"]').click();
     cy.get('[data-test="save-edit-button"]').should('not.exist');
     cy.get('.rosalution-logo').click();
     cy.get('.analysis-card').first().click();
-    cy.get(':nth-child(6) > .values > [data-test="value-row"]').should('contain', 'test');
-    cy.get('#Brief > tbody > :nth-child(4) > .values > [data-test="value-row"]')
-        .should('contain', 'Dr. Person One');
+    cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content > [data-test="value-row"]')
+      .should('contain', 'test');
+    cy.get('#Brief > div > [data-test="Nominator"] > .section-content > [data-test="value-row"]')
+      .should('contain', 'Dr. Person One');
   });
 });
