### 2023-07-01 #5443 @JDRomano2
@SeriousHorncat I agree with @suragnair - I've been struggling to get the software running on my local computer. I think two things would significantly help:

1. Provide troubleshooting details for deployment
2. Stand up a public demo server that will allow users to quickly explore the core functionality of the software.

To provide a bit more information about the issue I'm experiencing, after following all of the deployment instructions I'm seeing a `404 page not found` message when I navigate to `localhost:80` in my browser.

Nothing in the logs suggests the cause of the issue. E.g., for `rosalution-frontend-1`:
```
2023-07-01 09:11:07 yarn run v1.22.5
2023-07-01 09:11:07 $ vite --host 0.0.0.0 --port 80
2023-07-01 09:11:07
2023-07-01 09:11:07   VITE v4.0.4  ready in 241 ms
2023-07-01 09:11:07
2023-07-01 09:11:07   ➜  Local:   http://localhost:80/
2023-07-01 09:11:07   ➜  Network: http://172.18.0.2:80/
```

and for `rosalution-backend-1`:
```
2023-07-01 09:11:08 INFO:     Will watch for changes in these directories: ['/app']
2023-07-01 09:11:08 INFO:     Uvicorn running on http://0.0.0.0:8000 (Press CTRL+C to quit)
2023-07-01 09:11:08 INFO:     Started reloader process [11] using WatchFiles
2023-07-01 09:11:08 INFO:     Started server process [13]
2023-07-01 09:11:08 INFO:     Waiting for application startup.
2023-07-01 09:11:08 INFO:     Application startup complete.
```
