### 2023-07-01 #5443 @JDRomano2
Please note that I am satisfied with all other aspects of the study. If the authors are able to provide a usable example and improve documentation such that I can get everything running locally I should be able to assess the outstanding items in the checklist.

I do feel that this is within the scope of JOSS and is a valuable contribution to the field.
