### 2023-07-03 7e4a280
backend documentation updated for running pytest with the ROSALUTION_KEY env var set manually
