### 2023-07-03 9067e2f
updated the backend readme notes under testing to be pytest notes
