### 2023-07-03 8b8c21f
Frontned  unit tests updated along with additional console logging on frontend removed for debugging
