+++ frontend/test/components/AnalysisView/GeneBox.spec.js
@@ -123,8 +123,8 @@ describe('GeneBox.vue', () => {
     const gene = wrapper.findComponent('[data-test=gene-route]');
 
     expect(gene.props('to').name).to.equal('annotation');
-    expect(gene.props('to').params.analysis_name).to.equal('CPAM0046');
-    expect(gene.props('to').params.gene).to.equal('LMNA');
+    expect(gene.props('to').props.analysis_name).to.equal('CPAM0046');
+    expect(gene.props('to').props.gene).to.equal('LMNA');
   });
 
   it('should route to annotations for a variant in the anlaysis', async () => {
@@ -132,8 +132,8 @@ describe('GeneBox.vue', () => {
     const variant = wrapper.findComponent('[data-test=variant-route]');
 
     expect(variant.props('to').name).to.equal('annotation');
-    expect(variant.props('to').params.analysis_name).to.equal('CPAM0046');
-    expect(variant.props('to').params.gene).to.equal('LMNA');
-    expect(variant.props('to').params.variant).to.equal('NM_170707.3:c.745C>T(p.R249W)');
+    expect(variant.props('to').props.analysis_name).to.equal('CPAM0046');
+    expect(variant.props('to').props.gene).to.equal('LMNA');
+    expect(variant.props('to').props.variant).to.equal('NM_170707.3:c.745C>T(p.R249W)');
   });
 });
