+++ frontend/test/views/AnalysisView.spec.js
@@ -79,9 +79,9 @@ async function getMockedWrapper(latestStatus, mockedData) {
 
 describe('AnalysisView', () => {
   let mockedData;
-  let pedigreeAttachMock;
-  let pedigreeUpdateMock;
-  let pedigreeRemoveMock;
+  let attachSectionImageMock;
+  let updateSectionImageMock;
+  let removeSectionImageMock;
   let mockedAttachSupportingEvidence;
   let mockedRemoveSupportingEvidence;
   let mockedAttachThirdPartyLink;
@@ -96,9 +96,9 @@ describe('AnalysisView', () => {
     mockedData = sandbox.stub(Analyses, 'getAnalysis');
     mockedData.returns(fixtureData());
 
-    pedigreeAttachMock = sandbox.stub(Analyses, 'attachSectionImage');
-    pedigreeUpdateMock = sandbox.stub(Analyses, 'updateSectionImage');
-    pedigreeRemoveMock = sandbox.stub(Analyses, 'removeSectionImage');
+    attachSectionImageMock = sandbox.stub(Analyses, 'attachSectionImage');
+    updateSectionImageMock = sandbox.stub(Analyses, 'updateSectionImage');
+    removeSectionImageMock = sandbox.stub(Analyses, 'removeSectionImage');
 
     mockedAttachSupportingEvidence = sandbox.stub(Analyses, 'attachSupportingEvidence');
     mockedRemoveSupportingEvidence = sandbox.stub(Analyses, 'removeSupportingEvidence');
@@ -170,7 +170,7 @@ describe('AnalysisView', () => {
       try {
         await triggerAction(wrapper, 'Mark Ready');
       } catch (error) {
-        console.log(error);
+        console.error(error);
       }
       expect(toast.state.active).to.be.true;
       expect(toast.state.type).to.equal('error');
@@ -411,15 +411,11 @@ describe('AnalysisView', () => {
     describe('when an image section does not have an image', () => {
       it('accepts an image render as content', async () => {
         const newPedigreeSection = {
-          header: 'Pedigree',
-          content: [
-            {
-              field: 'image',
-              value: ['fakeimagefileid'],
-            },
-          ],
+          section: 'Pedigree',
+          field: 'Pedigree',
+          image_id: '64a2f06a4d4d29b8dc93c2d8',
         };
-        pedigreeAttachMock.returns(newPedigreeSection);
+        attachSectionImageMock.returns(newPedigreeSection);
 
         const pedigreeSection = wrapper.findComponent('[id=Pedigree]');
         pedigreeSection.vm.$emit('attach-image', 'Pedigree');
@@ -443,7 +439,12 @@ describe('AnalysisView', () => {
       beforeEach(() => {
         const imageSection = {
           header: 'Pedigree',
-          content: [{field: 'image', value: ['635a89aea7b2f21802b74539']}],
+          attachment_field: 'Pedigree',
+          content: [{
+            type: 'images-dataset',
+            field: 'Pedigree',
+            value: [{file_id: '635a89aea7b2f21802b74539'}],
+          }],
         };
         const analysisWithNewEvidence = fixtureData();
         const pedigreeSectionIndex =
@@ -454,36 +455,40 @@ describe('AnalysisView', () => {
       });
 
       it('updates section image content with input dialog', async () => {
-        pedigreeUpdateMock.returns({
-          header: 'Pedigree',
-          content: [{field: 'image', value: ['different-image-635a89aea7b2f21802b74539']}],
+        updateSectionImageMock.returns({
+          section: 'Pedigree',
+          field: 'Pedigree',
+          image_id: 'different-image-635a89aea7b2f21802b74539',
         });
 
[…]
