### 2023-07-03 8c96c96
Updated test_analysis_collection.py to reflect the new tests for the analysis section images. Also removed a few functions from analysis_collection.py pretaining to pedigree. Lastly, updated some text fixtures to more accurately reflect what the structure is
