+++ backend/tests/unit/repository/test_analysis_collection.py
@@ -133,16 +133,6 @@ def test_attach_file_supporting_evidence(analysis_collection, cpam0002_analysis_
     assert 'attachment_id' in new_evidence
     assert new_evidence['attachment_id'] == 'Fake-Mongo-Object-ID-2'
 
-
-def test_add_pedigree_file(analysis_collection, empty_pedigree):
-    """Tests adding pedigree file to an analysis and return an updated analysis"""
-    analysis_collection.collection.find_one.return_value = empty_pedigree
-    expected = read_test_fixture("analysis-CPAM0002.json")
-
-    analysis_collection.add_pedigree_file("CPAM0002", "63505be22888347cf1c275db")
-    analysis_collection.collection.find_one_and_update.assert_called_with({"name": "CPAM0002"}, {"$set": expected})
-
-
 def test_get_genomic_units(analysis_collection):
     """Tests the get_genomic_units function"""
     analysis_collection.collection.find_one.return_value = read_test_fixture("analysis-CPAM0002.json")
@@ -189,55 +179,49 @@ def test_remove_supporting_evidence(analysis_collection):
     actual = analysis_collection.remove_supporting_evidence("CPAM0002", "633afb87fb250a6ea1569555")
     assert actual == expected
 
+# HERE ARE THE PEDIGREE TESTS
 
-def test_get_pedigree_file_id(analysis_collection):
-    """Tests the get_pedigree_file_id function"""
-    analysis_collection.collection.find_one.return_value = read_test_fixture("analysis-CPAM0002.json")
-    actual = analysis_collection.get_pedigree_file_id("CPAM0002")
-    assert actual == "63505be22888347cf1c275db"
-
-
-def test_get_pedigree_file_id_pedigree_empty(analysis_collection, empty_pedigree):
-    """Tests the get_pedigree_file_id function"""
+def test_add_image_to_pedigree_section(analysis_collection, empty_pedigree):
+    """Tests adding an image to the pedigree section of the CPAM0002 analysis"""
     analysis_collection.collection.find_one.return_value = empty_pedigree
-    try:
-        analysis_collection.get_pedigree_file_id("CPAM0002")
-    except ValueError as error:
-        assert isinstance(error, ValueError)
-        assert str(error) == "Analysis CPAM0002 does not have a pedigree file"
-
-
-def test_get_pedigree_file_id_analysis_does_not_exist(analysis_collection):
-    """Tests the get_pedigree_file_id function"""
-    analysis_collection.collection.find_one.return_value = None
-    try:
-        analysis_collection.get_pedigree_file_id("CPAM2222")
-    except ValueError as error:
-        assert isinstance(error, ValueError)
-        assert str(error) == "Analysis with name CPAM2222 does not exist"
-
-
-def test_get_pedigree_file_id_pedigree_section_does_not_exist(analysis_collection):
-    """Tests the get_pedigree_file_id function"""
-    missing_pedigree_section = read_test_fixture("analysis-CPAM0002.json")
-    del missing_pedigree_section["sections"][2]
-    analysis_collection.collection.find_one.return_value = missing_pedigree_section
-    try:
-        analysis_collection.get_pedigree_file_id("CPAM0002")
-    except ValueError as error:
-        assert isinstance(error, ValueError)
-        assert str(error) == "Analysis CPAM0002 does not have a pedigree section"
-
-
-def test_remove_pedigree_file_id(analysis_collection):
-    """Tests the remove_pedigree_file_id function"""
-    analysis_collection.collection.find_one.return_value = read_test_fixture("analysis-CPAM0002.json")
     expected = read_test_fixture("analysis-CPAM0002.json")
-    expected["sections"][2]["content"] = []
[…]
