### 2023-07-03 11bea3d
Fixed the method that needing the logging to be patched causing an error with the unit tests
