### 2023-07-03 86c3325
Fixed up the integration tests for the analysis_router.py
