+++ backend/tests/integration/test_analysis_routers.py
@@ -4,6 +4,7 @@ import warnings
 import json
 import os
 import datetime
+from bson import ObjectId
 from unittest.mock import patch
 
 import pytest
@@ -50,43 +51,6 @@ def test_get_summary_by_name(client, mock_access_token, mock_repositories):
 
     assert response.status_code == 200
     assert response.json()["name"] == "CPAM0002"
-
-
-def test_import_analysis_from_phenotips_json(
-    client,
-    mock_access_token,
-    mock_repositories,
-    exported_phenotips_to_import_json,
-    mock_annotation_queue,
-    mock_security_get_current_user,  # pylint: disable=unused-argument
-):
-    """Testing if the create analysis endpoint creates a new analysis"""
-    mock_repositories["analysis"].collection.insert_one.return_value = True
-    mock_repositories["analysis"].collection.find_one.return_value = None
-    mock_repositories["genomic_unit"].collection.find_one.return_value = None
-    mock_repositories['genomic_unit'].collection.find.return_value = read_database_fixture("genomic-units.json")
-    mock_repositories['annotation_config'].collection.find.return_value = read_database_fixture(
-        "annotations-config.json"
-    )
-
-    with patch.object(BackgroundTasks, "add_task", return_value=None) as mock_background_add_task:
-        response = client.post(
-            "/analysis/import",
-            headers={"Authorization": "Bearer " + mock_access_token, "Content-Type": "application/json"},
-            json=exported_phenotips_to_import_json,
-        )
-
-        assert mock_annotation_queue.put.call_count == 29
-
-        mock_background_add_task.assert_called_once_with(
-            AnnotationService.process_tasks, mock_annotation_queue, mock_repositories['genomic_unit']
-        )
-
-    assert response.status_code == 200
-
-    response_data = json.loads(response.text)
-    assert response_data['latest_status'] == "Annotation"
-    assert response_data['timeline'][0]['username'] == 'johndoe'
 
 
 def test_import_analysis_with_phenotips_json(
@@ -172,41 +136,6 @@ def test_update_analysis_section(client, mock_access_token, mock_repositories, u
 #     )
 
 #     assert response
-
-
-def test_attach_pedigree_image(client, mock_access_token, mock_repositories):
-    """ Testing if the create analysis function works with file upload """
-    mock_repositories["analysis"].collection.find_one.return_value = read_test_fixture("analysis-CPAM0112.json")
-    expected = read_test_fixture("analysis-CPAM0112.json")
-    for section in expected["sections"]:
-        if section["header"] == "Pedigree":
-            section["content"] = [{'field': 'image', 'value': ["633afb87fb250a6ea1569555"]}]
-    mock_repositories['analysis'].collection.find_one_and_update.return_value = expected
-    mock_repositories['bucket'].bucket.put.return_value = "633afb87fb250a6ea1569555"
-
-    # This is used here because the 'read_fixture' returns a json dict rather than raw binary
-    # We actually want to send a binary file through the endpoint to simulate a file being sent
-    # then json.loads is used on the other end in the repository.
-    # This'll get updated and broken out in the test_utils in the future
-    path_to_current_file = os.path.realpath(__file__)
-    current_directory = os.path.split(path_to_current_file)[0]
-    path_to_file = os.path.join(current_directory, '../fixtures/' + 'pedigree-fake.jpg')
-
-    with open(path_to_file, "rb") as phenotips_file:
-        pedigree_image = phenotips_file.read()
-        pedigree_bytes = bytearray(pedigree_image)
-        response = client.post(
-            "/analysis/CPAM0112/attach/pedigree",
-            headers={"Authorization": "Bearer " + mock_access_token},
[…]
