### 2023-07-03 efd6dd4
Removed some comments
