### 2023-07-04 6584460
prototype for input forum validation
