### 2023-07-04 2fb8d6f
added the missing styling to main.css and updated the readme to include it
