### 2023-07-05 b159d8c
Updated the the annotation configuration to uitilize the tag dataset to visualize the HPO terms
