### 2023-07-05 a564554
progress in migration script
