+++ backend/src/routers/analysis_router.py
@@ -72,7 +72,7 @@ async def create_file(
 
     phenotips_importer = PhenotipsImporter(repositories["analysis"], repositories["genomic_unit"])
     try:
-        new_analysis = phenotips_importer.import_phenotips_json(phenotips_input)
+        new_analysis = phenotips_importer.import_phenotips_json(phenotips_input.dict())
         new_analysis['timeline'].append(Event.timestamp_create_event(username).dict())
         repositories['analysis'].create_analysis(new_analysis)
 
