### 2023-07-05 17db0c5
Added migration script to reconfigure analysis sections. Gene sections throwing error now
