+++ etc/fixtures/migrations/migrate-new-sections-phenotips-importer.js
@@ -7,7 +7,7 @@ mongosh /tmp/fixtures/migrations/migrate-new sections-phenotips-importer.js
     
     Example: 
     
-    mongosh --host localhost --port 27017 --eval "var help=true;" /tmp/fixtures/migrations/migrate-new-sections-phenotips-importer.js
+    mongosh --host localhost --port 27017 --eval 'var help=true;' /tmp/fixtures/migrations/migrate-new-sections-phenotips-importer.js
     
     docker exec -it <rosalution_db_container> mongosh /tmp/fixtures/migrations/migrate-new-sections-phenotips-importer.js
 `
@@ -18,9 +18,9 @@ if (help === true) {
 }
 
 if (typeof databaseName === 'undefined') {
-    databaseName = "rosalution_db";
+    databaseName = 'rosalution_db';
 } else if (typeof databaseName !== 'string') {
-    print("databaseName must be a string");
+    print('databaseName must be a string');
     quit(1);
 }
 
@@ -28,20 +28,20 @@ if (typeof databaseName === 'undefined') {
 // can we just use the config in phenotips importer? without creating an array/object here?
 // const newSectionsRename = [
 //   {
-//       "header": str(phenotips_json_data["gene"] + "Gene to Phenotype (like pedigree)"), "content":[]
+//       'header': str(phenotips_json_data['gene'] + 'Gene to Phenotype (like pedigree)'), 'content':[]
 //   }, {
-//       "header": str(phenotips_json_data["gene"] + "Molecular Mechanism"), "content":[
-//           {"field": 'Function', "value": []}
+//       'header': str(phenotips_json_data['gene'] + 'Molecular Mechanism'), 'content':[
+//           {'field': 'Function', 'value': []}
 //       ]
 //   },{
-//       "header": str(phenotips_json_data["gene"] + "Function"), "content":[]
+//       'header': str(phenotips_json_data['gene'] + 'Function'), 'content':[]
 //   },{
-//       "header": 'Model Goals', "content": [
-//           {"field": 'Model of Interest', "value": []},
-//           {"field": 'Goals', "value": []},
-//           {"field": 'Proposed Model/Project', "value": []},
-//           {"field": 'Existing Collaborations', "value": []},
-//           {"field": 'Existing Funding', "value": []},
+//       'header': 'Model Goals', 'content': [
+//           {'field': 'Model of Interest', 'value': []},
+//           {'field': 'Goals', 'value': []},
+//           {'field': 'Proposed Model/Project', 'value': []},
+//           {'field': 'Existing Collaborations', 'value': []},
+//           {'field': 'Existing Funding', 'value': []},
 //     ]
 //   }
 // ]
@@ -61,24 +61,88 @@ try {
         genes.push(unit.gene)
       }
     })
-    print(genes)
 
-    //iterating through the genes array to append the gene sections here
-    genes.forEach(gene => {
-      print ("APPENDING GENE SECTIONS HERE TO element.sections")
-    })
+    //iterating through the genes array to create the gene sections here
+    if (genes.length > 0) {
+      genes.forEach(gene => {
+        print (`${gene} APPENDING GENE SECTIONS HERE TO element.sections`)
+        newGeneSections = [{
+            "header": gene.concat(" Gene to Phenotype"), "content": [
+                {"type": "images-dataset", "field": 'Gene to Phenotype', "value": []},
+                {
+                    "type": "section-text", "field": 'HPO Terms',
+                    // move HPO TERMS HERE HOW???
+                    // Need to get it from Clinical History
+                    "value": "HPO TERMS TO BE MOVED HERE FROM CLINICAL HISTORY"
+                },
+            ]
+        }, {
[…]
