### 2023-07-06 91ac2fa
Updated gene sections to reconfigure where HPO terms are displayed
