### 2023-07-06 3d4732a
added the pedigree image if it exists
