### 2023-07-06 8c6bb6f
Removed print statements and included pedgree image migration
