### 2023-07-06 4575960
script fix, verified can only run it once succesfully
