+++ backend/src/core/phenotips_importer.py
@@ -116,7 +116,7 @@ class PhenotipsImporter:
             if genomic_unit['gene']:
                 new_sections = [{
                     "header": str(genomic_unit["gene"] + " Gene to Phenotype"), "content": [
-                        {"type": "images-dataset", "field": 'Gene to Phenotype', "value": []},
+                        {"type": "images-dataset", "field": str(genomic_unit["gene"] + " Gene to Phenotype"), "value": []},
                         {
                             "type": "section-text", "field": 'HPO Terms',
                             "value": [self.extract_hpo_terms(phenotips_json_data["features"])]
@@ -124,10 +124,10 @@ class PhenotipsImporter:
                     ]
                 }, {
                     "header": str(genomic_unit["gene"] + " Molecular Mechanism"),
-                    "content": [{"type": "section-text", "field": 'Function Overview', "value": []}]
+                    "content": [{"type": "section-text", "field": str(genomic_unit["gene"] + " Molecular Mechanism"), "value": []}]
                 }, {
                     "header": str(genomic_unit["gene"] + " Function"),
-                    "content": [{"type": "images-dataset", "field": 'Function', "value": []},]
+                    "content": [{"type": "images-dataset", "field": str(genomic_unit["gene"] + " Function"), "value": []},]
                 }]
                 analysis_data['sections'].extend(new_sections)
 
