### 2023-07-07 6945d16
Added some logging to the script to be informative of the output and fixed a few things found in the process
