### 2023-07-07 103c171
Phenotips importer update and corresponding Migration Script (#102)
* Added new sections to analysis

* Modified phenotips importer to show sections for multiple genes in analysis

* progress in migration script

* Fixed backend linting

* Fixed up phenotips_importer.py and added a test

* Added migration script to reconfigure analysis sections. Gene sections throwing error now

* Updated the CSS for the section text so that it wraps and doesn't go beyhond the bounds of its container.

* Updated gene sections to reconfigure where HPO terms are displayed

* Removed duplicate Clinical History/HP
[…]
