+++ backend/src/core/phenotips_importer.py
@@ -14,8 +14,7 @@ class PhenotipsImporter:
 
     def import_phenotips_json(self, phenotips_json_data):
         """Imports the phenotips json data into the database"""
-        if not isinstance(phenotips_json_data, dict):
-            phenotips_json_data = phenotips_json_data.dict()
+
         phenotips_variants = []
         variant_annotations = [
             "inheritance", "zygosity", "interpretation", "transcript", "cdna", "reference_genome", "protein", "gene"
@@ -63,7 +62,7 @@ class PhenotipsImporter:
         elif data_format == "gene":
             genomic_data = {"gene_symbol": data['gene'], "gene": data['gene'], "annotations": []}
         else:
-            warnings.warn("Invalid data format for import_genomic_unit_collection_data method")
+            warnings.warn("Invalid data format for import_genomic_unit_collection_data method", UserWarning)
             return None
         return genomic_data
 
@@ -74,24 +73,20 @@ class PhenotipsImporter:
             "name": str(phenotips_json_data["external_id"]).replace("-", ""), "description": "", "nominated_by": "",
             "genomic_units": [], "sections": [{
                 "header": 'Brief', "content": [
-                    {"field": 'Nominator', "value": []},
-                    {"field": 'Participant', "value": []},
-                    {"field": 'Phenotype', "value": []},
-                    {"field": 'Model of Interest', "value": []},
-                    {"field": 'Goals', "value": []},
-                    {"field": 'Proposed Model/Project', "value": []},
+                    {"type": "section-text", "field": 'Nominator', "value": []},
+                    {"type": "section-text", "field": 'Participant', "value": []},
+                    {"type": "section-text", "field": 'Phenotype', "value": []},
                 ]
             }, {
                 "header": 'Clinical History', "content": [
-                    {"field": 'Clinical Diagnosis', "value": []},
-                    {"field": 'Affected Individuals Identified', "value": []},
-                    {"field": 'Sequencing', "value": []},
-                    {"field": 'Testing', "value": []},
-                    {"field": 'Systems', "value": []},
-                    {"field": 'HPO Terms', "value": [self.extract_hpo_terms(phenotips_json_data["features"])]},
-                    {"field": 'Additional Details', "value": []},
+                    {"type": "section-text", "field": 'Clinical Diagnosis', "value": []},
+                    {"type": "section-text", "field": 'Affected Individuals Identified', "value": []},
+                    {"type": "section-text", "field": 'Sequencing', "value": []},
+                    {"type": "section-text", "field": 'Testing', "value": []},
+                    {"type": "section-text", "field": 'Systems', "value": []},
+                    {"type": "section-text", "field": 'Additional Details', "value": []},
                 ]
-            }, {"header": 'Pedigree', "content": []}]
+            }, {"header": 'Pedigree', "content": [{"type": "images-dataset", "field": "Pedigree", "value": []}]}]
         }
 
         for phenotips_gene in phenotips_genes:
@@ -116,6 +111,41 @@ class PhenotipsImporter:
                         analysis_unit['transcripts'].append(new_transcript)
 
             analysis_data['genomic_units'].append(analysis_unit)
+
+        for genomic_unit in analysis_data['genomic_units']:
+            if genomic_unit['gene']:
+                new_sections = [{
+                    "header": str(genomic_unit["gene"] + " Gene to Phenotype"), "content": [
+                        {
+                            "type": "images-dataset", "field": str(genomic_unit["gene"] + " Gene to Phenotype"),
+                            "value": []
+                        },
+                        {
+                            "type": "section-text", "field": 'HPO Terms',
+                            "value": [self.extract_hpo_terms(phenotips_json_data["features"])]
+                        },
+                    ]
+                }, {
[…]
