+++ system-tests/e2e/edit_case_analysis.cy.js
@@ -8,25 +8,19 @@ describe('edit_case_analysis.cy.js', () => {
     cy.get('.grey-rounded-menu').invoke('attr', 'style', 'display: block; visibility: hidden; opacity: 0;');
   });
 
-  // it('Enables editing an analysis', () => {
-  //   cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content > [data-test="editable-value"]')
-  //     .should('exist');
-  //   cy.get('[data-test="save-edit-button"]').should('exist');
-  // });
-
   it('edits a value & saves to the analysis', () => {
-    cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content > [data-test="editable-value"]')
+    cy.get('#Brief > div > [data-test="Phenotype"] > .section-content > [data-test="editable-value"]')
       .click();
-    cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content > [data-test="editable-value"]')
+    cy.get('#Brief > div > [data-test="Phenotype"] > .section-content > [data-test="editable-value"]')
       .type('test');
-    cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content > [data-test="editable-value"]')
+    cy.get('#Brief > div > [data-test="Phenotype"] > .section-content > [data-test="editable-value"]')
       .should('contain', 'test');
     cy.get('[data-test="save-edit-button"]').should('exist');
     cy.get('[data-test="save-edit-button"]').click();
     cy.get('[data-test="save-edit-button"]').should('not.exist');
     cy.get('.rosalution-logo').click();
     cy.get('.analysis-card').first().click();
-    cy.get('#Brief > div > [data-test="Proposed Model/Project"] > .section-content > [data-test="value-row"]')
+    cy.get('#Brief > div > [data-test="Phenotype"] > .section-content > [data-test="value-row"]')
       .should('contain', 'test');
     cy.get('#Brief > div > [data-test="Nominator"] > .section-content > [data-test="value-row"]')
       .should('contain', 'Dr. Person One');
