### 2023-07-07 43cd2ac
Updating system test to match the miggrated fixtures
