### 2023-07-07 ad958d8
Analysis Sections Enhancement with Migration and Additional Section Images Attachments (#103)
* wip to refactoring to supporting multiple sections in an analysis to add attachments

* wip for analysis backend integration with the adding/updating/removing

* debugging whats going on

* got it working?

* reduced the CSS footprint for the gene box component and the dom to be simpler

* wip for consolidating the css

* conslidated work that was here

* wip for refactoring the backend

* fixed frontend linting issues

* Refactoring to make the emits work, starting work on the updating section imag
[…]
