+++ backend/src/models/analysis.py
@@ -22,10 +22,11 @@ class GenomicUnit(BaseModel):
     variants: List = []
 
 
-class Section(BaseModel):
+class Section(BaseModel, frozen=True):
     """The sections of case notes associated with an analysis"""
 
     header: str
+    attachment_field: Optional[str] = None
     content: List = []
 
 

[…]
