### 2023-07-07 0c64771
Skipping the 'should allow the user to navigate the analysis via the logo, header, and section anchor links' for now until we fix it
