### 2023-07-07 0ef93e7
Annotation rendering configuration new section updates (#104)
* annotation rendering configuration frontend updates (#84)

* updated annotation configuration, updated unit tests, partially updated system tests

* updating the sidebar style to remained a fixed size, updated the logic for removing '_' from the anchors when displaying them

* fixed system test & made the anchors have consistent naming style

* added back the dataset data for each model

* lint fix

* unit test update

* removed file upload where not needed

* made requested PR changes, updating variant to hgvs_variant for attachm
[…]
