+++ backend/tests/integration/test_analysis_routers.py
@@ -87,7 +87,7 @@ def test_import_analysis_with_phenotips_json(
 
             phenotips_file.close()
 
-            assert mock_annotation_queue.put.call_count == 29
+            assert mock_annotation_queue.put.call_count == 45
 
             mock_background_add_task.assert_called_once_with(
                 AnnotationService.process_tasks, mock_annotation_queue, mock_repositories['genomic_unit']
