### 2023-07-07 uab-cgds-worthey/rosalution#87 @SeriousHorncat
Some applications have overlapping aspects; however, none address the direct problem area Rosalution aims to tackle.  We are working on updating the paper to mention some of these overlapping tools more explicitly.

We didn't include an explicit section for "State of the field' since we observed very few of the recent JOSS publications explicitly had this section in the papers.  Would more clearly defining the state of the field in the "Statement of Need" suffice?
