### 2023-07-07 uab-cgds-worthey/rosalution#105 @SeriousHorncat
[Bug]: Documentation for debugging deployment issues and improved instructions for visiting the application
### First Check

- [X] I added a very descriptive title to this issue.
- [X] Check that there isn't [already an issue](https://github.com/uab-cgds-worthey/rosalution/issues/) that reports the same bug to avoid creating a duplicate.
- [X] Follow our [Code of Conduct](https://github.com/uab-cgds-worthey/rosalution/blob/main/CODE_OF_CONDUCT.md)
- [X] Read the [Contributing Guidelines](https://github.com/uab-cgds-worthey/rosalution/blob/main/CONTRIBUTING.md).

### Describe the bug

I am using the main README.md to deploy Rosalution and am encountering several deployment issues.  I could utilize additional insight into how best to resolve these deployment issues.

### What browsers are you seeing the problem on if accessed via the browser?

_No response_

### Version

Chrome Version 114.0.5735.199 (Official Build) (64-bit)

### Steps to reproduce

_No response_

### Relevant log output

_No response_

### Code of Conduct

- [X] I agree to follow this project's Code of Conduct
