### 2023-07-07 #5443 @SeriousHorncat
> @SeriousHorncat would it be possible to include either a link to a demo website (pre-populated with a few examples) or a video covering basic usage in the GitHub README? This would allow potential users to get a quick overview of the functionality and how to use Rosalution.

@suragnair I meant to follow up on this sooner, we appreciate your patience. We agree a video demonstrating the basic usage of the application would be extremely helpful.  @lizworthey-cgds and are reviewing a case that would be a good example to use in a demonstration video. We will follow up once we have a demo video prepared.

> @SeriousHorncat I agree with @suragnair - I've been struggling to get the software running on my local computer. I think two things would significantly help:
>
> 1. Provide troubleshooting details for deployment
> 2. Stand up a public demo server that will allow users to quickly explore the core functionality of the software.
>
> To provide a bit more information about the issue I'm experiencing, after following all of the deployment instructions I'm seeing a `404 page not found` message when I navigate to `localhost:80` in my browser.
>
> Nothing in the logs suggests the cause of the issue. E.g., for `rosalution-frontend-1`:
>
> ```
> 2023-07-01 09:11:07 yarn run v1.22.5
> 2023-07-01 09:11:07 $ vite --host 0.0.0.0 --port 80
> 2023-07-01 09:11:07
> 2023-07-01 09:11:07   VITE v4.0.4  ready in 241 ms
> 2023-07-01 09:11:07
> 2023-07-01 09:11:07   ➜  Local:   http://localhost:80/
> 2023-07-01 09:11:07   ➜  Network: http://172.18.0.2:80/
> ```
>
> and for `rosalution-backend-1`:
>
> ```
> 2023-07-01 09:11:08 INFO:     Will watch for changes in these directories: ['/app']
> 2023-07-01 09:11:08 INFO:     Uvicorn running on http://0.0.0.0:8000 (Press CTRL+C to quit)
> 2023-07-01 09:11:08 INFO:     Started reloader process [11] using WatchFiles
> 2023-07-01 09:11:08 INFO:     Started server process [13]
> 2023-07-01 09:11:08 INFO:     Waiting for application startup.
> 2023-07-01 09:11:08 INFO:     Application startup complete.
> ```

@JDRomano2
1.  Additional and clearer documentation on how to visit the application, along with additional documentation on debugging common deployment issues, would be incredibly helpful.  I will create an issue in our repo to track this.  https://github.com/uab-cgds-worthey/rosalution/issues/105

2.  We are discussing amongst the maintainers and our PI the potential for a demo site.  We would need to remove certain features such as uploading new cases; since that could potentially put a strain on the 3rd party APIs. We try to be careful and limit the potential vectors of misuse of the application.

In the meantime, lets try to help get y'all deploying the application on your machines.  We want to help continue to remove barriers for future contributors.

After running the following docker compose command in the code block below, could you visit `http://local.rosalution.cgds/rosalution/` and the `http://local.rosalution.cgds/rosalution/api/docs` instead of visiting  `localhost:80` ?

```bash
docker compose up --build
```

If this does not work, could you share the results of the `./setup.sh` and the `docker ps` commands from below?

```bash
./setup.sh clean
docker compose down
docker system prune -a --volumes # This will also remove other docker images but it's the fastest way to clear old images
docker compose up
docker ps
```

note: run the below code block in the project root folder of Rosalution
