### 2023-07-11 91f9058
Updated the README to discuss how to visit the locally deployed application  on your computer after running docker compose
