### 2023-07-20 8df4bd3
Created migration script to fixed the field names to be consistent and to add attachment_field the missing sections
