### 2023-07-20 f39895c
Changed backend-logging to use the logger rather than a file
