### 2023-07-20 1fd8d71
Created migration script to fixed the field names to be consistent an… (#107)
* Created migration script to fixed the field names to be consistent and to add attachment_field the missing sections
