### 2023-07-21 3b17714
Updated the diectory for logging.conf and changed the formatting for annotation to be more readable
