### 2023-07-26 f109502
Updated the logging to have its own annotation_logging file that formats the logging properly
