### 2023-07-26 9f49666
Updated the readme for logging in the backend
