### 2023-07-26 15fda1d
Replaced existing print statements with loggers, updated wording on annotation finalization, and added more documentation
