+++ README.md
@@ -429,7 +429,6 @@ To report a bug, refer to [🐞Reporting Issues](CONTRIBUTING.md#reporting-issue
     - Email: aeunoantonison@uabmc.edu
 - [Rabab Fatima](https://github.com/fatimarabab)
 - [James Scherer](https://github.com/JmScherer)
-- [Alex Moss](https://github.com/kuthedk)
 
 ---
 
