### 2023-07-27 ebbf1ab
Backend logging (#108)
* logging working

* More logging stuff

* Changed backend-logging to use the logger rather than a file

* Updated the diectory for logging.conf and changed the formatting for annotation to be more readable

* Some linting and backend tests passing

* Adding in some preliminary comments

* Updated the logging to have its own annotation_logging file that formats the logging properly

* Should be passing now

* Updated the readme for logging in the backend

* Replaced existing print statements with loggers, updated wording on annotation finalization, and added more documen
[…]
