+++ backend/tests/unit/core/test_annotate.py
@@ -14,21 +14,15 @@ def test_queuing_annotations_for_genomic_units(cpam0046_analysis, annotation_col
     assert mock_queue.put.call_count == 45
 
 
-# Patching the temporary helper method that is writing to a file, this will be
-# removed once that helper method is no longer needed for the development
-
-
 # The patched method sare done provided in reverse order within the test param arguments.  Was accidently getting
 # logging mock results instead of non task type annotations.  Was causing major failures in verifying values
-@patch("src.core.annotation.log_to_file")
 @patch("src.core.annotation_task.AnnotationTaskInterface.extract")
 @patch("src.core.annotation_task.ForgeAnnotationTask.annotate")
 @patch("src.core.annotation_task.HttpAnnotationTask.annotate")
 @patch("src.core.annotation_task.NoneAnnotationTask.annotate")
 def test_processing_cpam0046_annotation_tasks(
-    none_task_annotate, http_task_annotate, forge_task_annotate, annotate_extract_mock, log_to_file_mock,
-    cpam0046_annotation_queue
-):  # pylint: disable=unused-argument
+    none_task_annotate, http_task_annotate, forge_task_annotate, annotate_extract_mock, cpam0046_annotation_queue
+):
     """Verifies that each item on the annotation queue is read and executed"""
     flag = {'dependency_flag_passed': False}
 
@@ -59,7 +53,6 @@ def test_processing_cpam0046_annotation_tasks(
     assert annotate_extract_mock.call_count == 45
 
 
-@patch("src.core.annotation.log_to_file")
 @patch(
     "src.core.annotation_task.AnnotationTaskInterface.extract",
     return_value=[{
@@ -73,9 +66,8 @@ def test_processing_cpam0046_annotation_tasks(
 @patch("src.core.annotation_task.HttpAnnotationTask.annotate")
 @patch("src.core.annotation_task.NoneAnnotationTask.annotate")
 def test_processing_cpam0002_annotations_tasks(
-    none_task_annotate, http_task_annotate, forge_task_annotate, annotate_extract_mock, log_to_file_mock,
-    cpam0002_annotation_queue
-):  # pylint: disable=unused-argument
+    none_task_annotate, http_task_annotate, forge_task_annotate, annotate_extract_mock, cpam0002_annotation_queue
+):
     """
         Verifies that the annotation collection is being sent the proper amount of extracted annotations for
         CPAM analysis 0002
