### 2023-07-27 cda20c8
Working script for updating the genomic unit fixture in ./etc/initial-seed/' '
