### 2023-07-28 b800fa7
Almost done with the script, just committing
