### 2023-07-28 f321d9b
finalizing the script
