### 2023-07-31 2faf15f
Backend dockerfile now copies over the logging.conf for the production deployment
