+++ backend/Dockerfile
@@ -9,6 +9,7 @@ ENTRYPOINT ["/bin/sh", "-c", "uvicorn src.main:app --host 0.0.0.0 --port 8000 --
 # Production Build Stage
 FROM python:3.9.10 as production-stage
 WORKDIR /app
+COPY logging.conf /app/logging.conf
 COPY requirements.txt /app/requirements.txt
 RUN pip install --no-cache-dir -r /app/requirements.txt
 COPY ./src /app/src
