### 2023-08-07 6699bad
Adding a check for jq in the update-genomic-units-fixtures.sh script
