### 2023-08-07 c4b0778
Added configuration for DITTO dataset to show up in Annotation View. Corrected typos for Score Dataset.
