### 2023-08-07 1ffbc7f
Adding the ditto score script and ditto score csv
