### 2023-08-07 f3095c2
Adding docker run command
