### 2023-08-07 5873d79
Genomic unit fixture updater (#111)
* Working script for updating the genomic unit fixture in ./etc/initial-seed/'
'

* Almost done with the script, just committing

* finalizing the script

* Adding a check for jq in the update-genomic-units-fixtures.sh script
