### 2023-08-07 b99acfd
Fixed unit test with updated fixtures
