### 2023-08-07 46d1dba
DITTO Frontend Dataset Configuration (#113)
* Added configuration for DITTO dataset to show up in Annotation View. Corrected typos for Score Dataset.

* Fixed typo in system tests file, and skipped a test to be handled when future changes are made to the score dataset component.
