### 2023-08-07 83220e6
Ditto migration script (#115)
* Adding the ditto score script and ditto score csv

* Adding docker run command
