### 2023-08-10 d6ea9e1
Updated the write check, don't need a read check
