### 2023-08-11 a302da9
Added scope checks for all write only endpoints
