+++ backend/src/routers/analysis_router.py
@@ -69,9 +69,12 @@ async def create_file(
     phenotips_file: Union[bytes, None] = File(default=None),
     repositories=Depends(database),
     annotation_task_queue=Depends(annotation_queue),
-    username: VerifyUser = Security(get_current_user)
+    username: VerifyUser = Security(get_current_user),
+    authorized=Security(get_authorization, scopes=["write"])
 ):
     """ Imports a .json file for a phenotips case """
+    logger.info(authorized)
+
     # Quick and dirty json loads
     phenotips_input = BasePhenotips(**json.loads(phenotips_file))
 
@@ -93,8 +96,13 @@ async def create_file(
 
 
 @router.put("/{analysis_name}/mark_ready", response_model=Analysis)
-def mark_ready(analysis_name: str, repositories=Depends(database), username: VerifyUser = Security(get_current_user)):
+def mark_ready(
+    analysis_name: str, repositories=Depends(database),
+    username: VerifyUser = Security(get_current_user),
+    authorized=Security(get_authorization, scopes=["write"])
+):
     """ Marks an analysis as ready for review """
+    logger.info(authorized)
     try:
         return repositories["analysis"].mark_ready(analysis_name, username)
     except ValueError as exception:
@@ -102,8 +110,13 @@ def mark_ready(analysis_name: str, repositories=Depends(database), username: Ver
 
 
 @router.put("/{analysis_name}/update/sections", response_model=Analysis)
-def update_analysis_sections(analysis_name: str, updated_sections: dict, repositories=Depends(database)):
+def update_analysis_sections(
+    analysis_name: str,
+    updated_sections: dict, repositories=Depends(database),
+    authorized=Security(get_authorization, scopes=["write"])    
+):
     """Updates the sections that have changes"""
+    logger.info(authorized)
     for (header, field) in updated_sections.items():
         for (updated_field, value) in field.items():
             if "Nominator" == updated_field:
@@ -139,10 +152,12 @@ def attach_section_image(
     upload_file: UploadFile = File(...),
     section_name: str = Form(...),
     field_name: str = Form(...),
-    repositories=Depends(database)
+    repositories=Depends(database),
+    authorized=Security(get_authorization, scopes=["write"])
 ):
     """ Saves the uploaded image it to the specified field_name in the analysis's section."""
 
+    logger.info(authorized)
     try:
         new_file_object_id = repositories["bucket"].save_file(
             upload_file.file, upload_file.filename, upload_file.content_type
@@ -164,10 +179,12 @@ def update_analysis_section_image(
     upload_file: UploadFile = File(...),
     section_name: str = Form(...),
     field_name: str = Form(...),
-    repositories=Depends(database)
+    repositories=Depends(database),
+    authorized=Security(get_authorization, scopes=["write"])
 ):
     """ Replaces the existing image by the file identifier with the uploaded one. """
 
+    logger.info(authorized)
     # This needs try catch like in annotation router
     new_file_id = repositories["bucket"].save_file(upload_file.file, upload_file.filename, upload_file.content_type)
 
@@ -182,9 +199,11 @@ def remove_analysis_section_image(
     file_id: str,
     section_name: str = Form(...),
     field_name: str = Form(...),
-    repositories=Depends(database)
+    repositories=Depends(database),
+    authorized=Security(get_authorization, scopes=["write"])
 ):
     """ Removes the image from an analysis section's field by its file_id """
+    logger.info(authorized)
     try:
         repositories['analysis'].remove_analysis_section_file(analysis_name, section_name, field_name, file_id)
     except Exception as exception:
@@ -231,8 +250,10 @@ def attach_third_party_link(
     third_party_enum: ThirdPartyLinkType,
[…]
