### 2023-08-11 d43642d
Resolved pylint issues for now. Need to refactor later.
