+++ backend/src/routers/analysis_router.py
@@ -22,19 +22,14 @@ from ..models.phenotips_json import BasePhenotips
 from ..models.user import VerifyUser
 from ..security.security import get_authorization, get_current_user
 
-# This is temporarily changed as security is removed for the analysis endpoints to make development easier
-# dependencies=[Depends(database), Security(get_authorization, scopes=["write"])]
-# and add the following dependencies at the top:
-
 logger = logging.getLogger(__name__)
 
 router = APIRouter(prefix="/analysis", tags=["analysis"], dependencies=[Depends(database)])
 
 
 @router.get("/", response_model=List[Analysis])
-def get_all_analyses(repositories=Depends(database), authorized=Security(get_authorization, scopes=["write"])):
+def get_all_analyses(repositories=Depends(database)):
     """Returns every analysis available"""
-    logger.info(authorized)
     return repositories["analysis"].all()
 
 
