### 2023-08-11 1fab08d
Backend security lockdown (#118)
* initial commit

* Updated the write check, don't need a read check

* Added scope checks for all write only endpoints

* Resolved pylint issues for now. Need to refactor later.

* Fixed formatting issues.

* Removed comments, and example code.

* fixed tests.

---------

Co-authored-by: James Scherer <jscherer@uab.edu>
