+++ backend/src/routers/analysis_router.py
@@ -1,4 +1,8 @@
+# pylint: disable=too-many-arguments
+# Due to adding scope checks, it's adding too many arguments (7/6) to functions, so diabling this for now.
+# Need to refactor later.
 """ Analysis endpoint routes that serve up information regarding anaysis cases for rosalution """
+import logging
 import json
 
 from typing import List, Optional, Union
@@ -16,13 +20,11 @@ from ..models.event import Event
 from ..enums import ThirdPartyLinkType
 from ..models.phenotips_json import BasePhenotips
 from ..models.user import VerifyUser
-from ..security.security import get_current_user
+from ..security.security import get_authorization, get_current_user
 
-# This is temporarily changed as security is removed for the analysis endpoints to make development easier
-# dependencies=[Depends(database), Security(get_authorization, scopes=["write"])]
-# and add the following dependencies at the top:
-# from ..security.security import get_authorization
-router = APIRouter(prefix="/analysis", tags=["analysis"], dependencies=[Depends(database), Security(get_current_user)])
+logger = logging.getLogger(__name__)
+
+router = APIRouter(prefix="/analysis", tags=["analysis"], dependencies=[Depends(database)])
 
 
 @router.get("/", response_model=List[Analysis])
@@ -64,9 +66,12 @@ async def create_file(
     phenotips_file: Union[bytes, None] = File(default=None),
     repositories=Depends(database),
     annotation_task_queue=Depends(annotation_queue),
-    username: VerifyUser = Security(get_current_user)
+    username: VerifyUser = Security(get_current_user),
+    authorized=Security(get_authorization, scopes=["write"])
 ):
     """ Imports a .json file for a phenotips case """
+    logger.info(authorized)
+
     # Quick and dirty json loads
     phenotips_input = BasePhenotips(**json.loads(phenotips_file))
 
@@ -88,8 +93,14 @@ async def create_file(
 
 
 @router.put("/{analysis_name}/mark_ready", response_model=Analysis)
-def mark_ready(analysis_name: str, repositories=Depends(database), username: VerifyUser = Security(get_current_user)):
+def mark_ready(
+    analysis_name: str,
+    repositories=Depends(database),
+    username: VerifyUser = Security(get_current_user),
+    authorized=Security(get_authorization, scopes=["write"])
+):
     """ Marks an analysis as ready for review """
+    logger.info(authorized)
     try:
         return repositories["analysis"].mark_ready(analysis_name, username)
     except ValueError as exception:
@@ -97,8 +108,14 @@ def mark_ready(analysis_name: str, repositories=Depends(database), username: Ver
 
 
 @router.put("/{analysis_name}/update/sections", response_model=Analysis)
-def update_analysis_sections(analysis_name: str, updated_sections: dict, repositories=Depends(database)):
+def update_analysis_sections(
+    analysis_name: str,
+    updated_sections: dict,
+    repositories=Depends(database),
+    authorized=Security(get_authorization, scopes=["write"])
+):
     """Updates the sections that have changes"""
+    logger.info(authorized)
     for (header, field) in updated_sections.items():
         for (updated_field, value) in field.items():
             if "Nominator" == updated_field:
@@ -134,10 +151,12 @@ def attach_section_image(
     upload_file: UploadFile = File(...),
     section_name: str = Form(...),
     field_name: str = Form(...),
-    repositories=Depends(database)
+    repositories=Depends(database),
+    authorized=Security(get_authorization, scopes=["write"])
 ):
     """ Saves the uploaded image it to the specified field_name in the analysis's section."""
 
+    logger.info(authorized)
     try:
         new_file_object_id = repositories["bucket"].save_file(
[…]
