### 2023-08-11 f5788c5
Removed the extra user01 and user03 as they are not needed for system tests anymore and revised them to use users that have usernames that are more clear for their roles
