### 2023-08-14 06b25bd
Apply suggestions from code review
Co-authored-by: James Scherer <jscherer@uab.edu>
Signed-off-by: Angelina Uno-Antonison <angelina.antonison@gmail.com>
