### 2023-09-04 uab-cgds-worthey/rosalution#87 @suragnair
Yes, defining the state of the field in the "Statement of Need" would be great!
