### 2023-09-04 #5443 @suragnair
Sorry for the delay! I was able to successfully install and get Rosalution running. The only items that remain are:
- A researcher friendly user guide or video that @SeriousHorncat mentioned they are preparing
- a comment on the state of the field in the manuscript
