### 2023-09-12 #5443 @SeriousHorncat
> Sorry for the delay! I was able to successfully install and get Rosalution running. The only items that remain are:
>
> * A researcher friendly user guide or video that @SeriousHorncat mentioned they are preparing
  * This has been perfect timing.  We have finished preparing our first pass for a demonstration video of how Rosalution is being currently used - https://uab.app.box.com/s/r76kbgsmu2w562akfvdcfhutu2jwe6u3 , we have updated our README.md to include a link out to the demonstration video.
> * a comment on the state of the field in the manuscript
  * Originally we planed to update the *Statement of Need* section, but it just kept growing, so we opted to add a whole section dedicated to the *State of the field* https://github.com/uab-cgds-worthey/rosalution/pull/134  https://github.com/uab-cgds-worthey/rosalution/issues/87.  So these changes are included now

Please let us know if there is anything else we can do to help facilitate this review more.
