### 2023-09-13 #5443 @SeriousHorncat
@suragnair and @JDRomano2  Thank you so much for investing your time for reviewing Rosalution.  We at @uab-cgds-worthey are grateful for the feedback which helps reduce barriers for future contributors of Rosalution. Also, @jmschrei thank you for your time as editor to help this review process proceed smoothly.

Rosalution has a tag and release for version [0.6.0](https://github.com/uab-cgds-worthey/rosalution/releases/tag/0.6.0).  What are the next steps necessary to help move this submission forward?
