### 2023-09-26 #5443 @SeriousHorncat
> @SeriousHorncat I am the AEiC on this track and here to help process final steps. I've checked your repository, the archive link, this review, and the paper. Most seems in order. Can you check the below final points please:

* [x]  The title for the archive does not match the paper, please edit the archive title to match.
    * Have revised the archive to match the title of the paper
* [x]  Check spelling of author `Alexander Moss` in the paper vs `Alex Moss` for the archive. Can you make these the same?
    * Resolved the name abbreviation from Alex to Alexander for their author entry
* [x]  Check if the word/expression `simul` is correct in the paper.
    * I read over the passage and followed up with looking for the definition to be sure. https://www.dictionary.com/browse/simul, several of our contributors use British english, simul can be used to to described as together and also used in place of simultaneous in British english.  In the case of the paper is describing that datasets and analyses are being annotated together via a multi-threaded processing pool via different REST APIs.  Is it ok if we leave the usage as is?
> * [x]  Please add the country to each affiliation. And please spell out the USA as United States of America.
    * Followed up and updated the country to each affiliation in the Zenodo archive and updated the paper in the repository accordingly. I did decide to create a new release so that it may exactly match the archive since we package the paper with the version.

**New Version**: 0.6.1
**Updated Zenodo Archive for 0.6.1**: 10.5281/zenodo.8380641
