### 2023-10-02 #5443 @Kevin-Mattheus-Moerman
@SeriousHorncat congratulations on this paper in JOSS!

@jmschrei thanks for editing this work!

And a special thanks to the reviewers: @suragnair, @JDRomano2
